### 2020-06-20 #385 @agricolamz
Submission: phonfieldwork
Submitting Author: George Moroz (@agricolamz)
Repository: https://github.com/agricolamz/phonfieldwork
Version submitted: 0.0.8
Version accepted: 2020-10-20

---

```
Package: phonfieldwork
Title: Linguistic Phonetic Fieldwork Tools
Version: 0.0.8
    tuneR,
    phonTools,
    grDevices,
    graphics,
    xml2
Authors@R: c(person("George", "Moroz",
             email = "agricolamz@gmail.com",
             comment = c(ORCID="0000-0003-1990-6083")))
Description: There are a lot of different typical tasks that have to be solved during phonetic research and experiments. This includes creating a presentation that will contain all stimuli, renaming and concatenating multiple sound files recorded during a session, automatic annotation in 'Praat' TextGrids (this is one of the sound annotation standards provided by 'Praat' software, see Boersma & Weenink 2018 <http://www.fon.hum.uva.nl/praat/>), creating an html table with annotations and spectrograms, and converting multiple formats ('Praat' TextGrid, 'ELAN', 'EXMARaLDA', 'Audacity', subtitles '.srt', and 'FLEx' flextext). All of these tasks can be solved by a mixture of different tools (any programming language has programs for automatic renaming, and Praat contains scripts for concatenating and renaming files, etc.). 'phonfieldwork' provides a functionality that will make it easier to solve those tasks independently of any additional tools. You can also compare the functionality with other packages: 'rPraat' <https://CRAN.R-project.org/package=rPraat>, 'textgRid' <https://CRAN.R-project.org/package=textgRid>.
    covr,
    DT,
    lingtypology
```

	- [x] workflow automataion

phonfieldowrks is a tool that helps researchers create sound annotation viewer from multiple sounds. I believe that sound annotation viewer helps to make phonetic more reproducible, since it make sound data widely availible. It helps reducing the time that annotators spent during annotations in Praat, Elan and EXMARaLDA, so it partially replace its functionality.

Linguists, phoneticians. Probably in the future specialists from bioacustics.

There are packages 'rPraat' and 'textgRid' <https://CRAN.R-project.org/package=textgRid>, that overlap in functionality with my package, but my package has wider coverage of tasks and especially created for specific workflow described in the docs (https://agricolamz.github.io/phonfieldwork/).
