### 2020-06-23 #385 @annakrystalli
Hello @agricolamz and many thanks for your submission.

We are discussing whether the package is in scope and need a bit more information.

- If a package has overlapping functionality with other packages, we require it to demonstrate in the documentation [how it is best in class](https://devguide.ropensci.org/policies.html#overlap). Could you add a more detailed comparison to the packages you mention in the README so we can evaluate?
- The package is to be used with data from human subjects and also mentions the publication of data.  rOpenSci have specific policies on [Ethics, Data Privacy and Human Subjects Research](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research) so we would expect care to be incorporated into package workflows to protect personal data and adhere to what subjects have consented to. Could you elaborate on any steps the package takes to ensure it conforms to these policies? A separate vignette might at some point be useful but a section in the README addressing these issues will be fine for now.
