### 2020-06-24 #385 @agricolamz
Hello @annakrystalli!

thank you for your response.

### other packages
In the README file I've mentiond [`rPraat`](https://CRAN.R-project.org/package=rPraat) and [`textgRid`](https://CRAN.R-project.org/package=textgRid). The first one get an access to the functionality of one particular program for phonetic/phonological research `Praat` and its annotation system (`TextGrid`). The second one is limited to operations with `Praat`'s annotation system (`TextGrid`).

My package has another perspective. The main goal of the my package is to provide different tools for phonetic/phonological researcher. In order to do it I provided several functions that are not connected to `Praat`, but really important for the process of the phonetic/phonological research:

* stimuli presentation creation
* soundfile renaming and automatic merge
* automatic annotation (connected to Praat in a sense that the result is a Praat Textgrid)
* importing data from other linguistic annotation programs (Elan, EXMARaLDA and FieldWorks)
* cut sounds according to annotation
* create a sound viewer that simplifies research process and could be a standard for data publication
* I have plans to enrich my package with the simmilar video functionality in order to help Sign language researchers

Summing up: my package is focused on the process of a research. Other packages are focused on replication of the functionality of Praat in R.

### ethic policy

Honestly, I didn't thought about it. That doesn't mean that I don't care. This **do** mean that I need help on this topic and whatever path this submission will take, I will appreciate any help and ideas about it. I will try to list here what I know about the topic.

Simple linguistic research or documentation looks like this: researcher record some audio/video files of the speech events and then extract some linguistic features from it. Sometimes it is possible to conduct a research in a big cities with mostly literate population, then consent is not a problem. But there are also a hue amount of work that performed in a "field", where researches go to small settlements with mostly rural population. In most cases it is possible to receive a consent, but it is hard to explain what does it mean.

In some countries there are an ethic commission that decides whether it is possible to conduct a research, but in some countries there is no any policy on this question (e. g. in my own country, Russia).

Since my package make it easier to publish multiple speech samples that could be a problem, since people could be identified (especially if they are from the small community). What can I do? I can insert in my [documentation](https://agricolamz.github.io/phonfieldwork/) and the `create_viewer()` function notes about what kind of consent should be gathered in order to satisfy research standards. It is easy. I would be happy if somebody help me, but I think I can gather this from other linguists.

I think that there are several things that participant should agree on:

1) conduct scientific research on gathered data;
2) publication of gathered data;

If participant agrees only on the first one, it would be useful to create an access restriction (e. g. password protection) on generated html viewer. This will limit users to academic circles. Which `rmarkdown` password protection strategy is better for this purpose?

I hope, I explained my position and my ideas, but I hope that rOpenSci community can help to elaborate this problem.
