+++ R/eaf_to_df.R
@@ -4,8 +4,8 @@
 #'
 #' @author George Moroz <agricolamz@gmail.com>
 #'
-#' @param eaf string with a filename or path to the .eaf file
-#' @return a dataframe with columns:  \code{tier}, \code{id}, \code{content}, \code{tier_name}, \code{tier_type}, \code{ts_start}, \code{ts_end}, \code{a_id}, \code{ar}).
+#' @param file_name string with a filename or path to the .eaf file
+#' @return a dataframe with columns:  \code{tier}, \code{id}, \code{content}, \code{tier_name}, \code{tier_type}, \code{time_start}, \code{time_end}, \code{a_id}, \code{ar}).
 #'
 #' @examples
 #' eaf_to_df(system.file("extdata", "test.eaf", package = "phonfieldwork"))
@@ -17,9 +17,9 @@
 #' @importFrom xml2 xml_text
 #' @importFrom xml2 xml_children
 
-eaf_to_df <- function(eaf){
+eaf_to_df <- function(file_name){
   # read file
-  l <- xml2::read_xml(eaf)
+  l <- xml2::read_xml(file_name)
   # extract tiers
   t <- xml2::xml_find_all(l, "TIER")
   # extract tiers
