+++ man/flextext_to_df.Rd
@@ -4,10 +4,10 @@
 \alias{flextext_to_df}
 \title{FLEX's .flextext file to dataframe}
 \usage{
-flextext_to_df(flextext)
+flextext_to_df(file_name)
 }
 \arguments{
-\item{flextext}{string with a filename or path to the .flextext file}
+\item{file_name}{string with a filename or path to the .flextext file}
 }
 \value{
 a dataframe with columns: \code{s_id} (that has structure paragraph_id.phrase_id), \code{txt}, \code{cf}, \code{hn}, \code{gls}, \code{msa}, \code{morph}, \code{word}, \code{phrase}, \code{paragraph}, \code{free_trans}, \code{text}, \code{text_title}
