+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 export(add_leading_symbols)
 export(annotate_textgrid)
+export(audacity_to_df)
 export(concatenate_soundfiles)
 export(create_glossed_document)
 export(create_image_look_up)
@@ -41,6 +42,7 @@ importFrom(tuneR,readMP3)
 importFrom(tuneR,readWave)
 importFrom(tuneR,writeWave)
 importFrom(utils,installed.packages)
+importFrom(utils,read.delim)
 importFrom(utils,write.csv)
 importFrom(xml2,read_xml)
 importFrom(xml2,xml_attr)
