### 2020-06-26 2a5fede
correct point tier visualisation
