+++ R/textgrid_to_df.R
@@ -7,7 +7,7 @@
 #' @param textgrid string with a filename or path to the TextGrid
 #' @param encoding TextGrid encoding. Import from \code{readLines()} function.
 #'
-#' @return a dataframe with columns:  \code{id}, \code{start}, \code{end} (if it is an interval tier -- the same as the start value), \code{annotation}, and \code{tier}
+#' @return a dataframe with columns:  \code{id}, \code{time_start}, \code{time_end} (if it is an interval tier -- the same as the start value), \code{annotation}, and \code{tier}
 #'
 #' @examples
 #' textgrid_to_df(example_textgrid)
@@ -24,16 +24,9 @@ textgrid_to_df <- function(textgrid, encoding = "unknown"){
   lapply(1:n_tiers, function(x){
     df <- phonfieldwork::tier_to_df(tg, x)
     df$tier <- x
-    if(!("end" %in% names(df))){
-      df <- data.frame(id = df$id,
-                       start = df$start,
-                       end = df$start,
-                       annotation = df$annotation,
-                       tier = df$tier)
-    }
     return(df)
   }) ->
     l
   result <- Reduce(rbind, l)
-  return(result[order(result$start),])
+  return(result[order(result$time_start),])
 }
