### 2020-06-26 9c1222e
Reduce() to do.call()
