### 2020-06-26 f764e60
add srt_to_df() function
