+++ man/srt_to_df.Rd
@@ -1,0 +1,26 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/srt_to_df.R
+\name{srt_to_df}
+\alias{srt_to_df}
+\title{Subtitles .srt file to dataframe}
+\usage{
+srt_to_df(file_name, encoding = "unknown")
+}
+\arguments{
+\item{file_name}{string with a filename or path to the .srt file}
+
+\item{encoding}{.srt encoding. Import from \code{readLines()} function.}
+}
+\value{
+a dataframe with columns:  \code{id}, \code{content}, \code{time_start}, \code{time_end}.
+}
+\description{
+Convert subtitles .srt file to a dataframe.
+}
+\examples{
+srt_to_df(system.file("extdata", "test.srt", package = "phonfieldwork"))
+
+}
+\author{
+George Moroz <agricolamz@gmail.com>
+}
