+++ NAMESPACE
@@ -22,6 +22,7 @@ export(get_textgrid_names)
 export(rename_soundfiles)
 export(rename_videofiles)
 export(set_textgrid_names)
+export(srt_to_df)
 export(textgrid_to_df)
 export(tier_to_df)
 importFrom(grDevices,dev.off)
