+++ man/eaf_to_df.Rd
@@ -10,7 +10,7 @@ eaf_to_df(file_name)
 \item{file_name}{string with a filename or path to the .eaf file}
 }
 \value{
-a dataframe with columns:  \code{tier}, \code{id}, \code{content}, \code{tier_name}, \code{tier_type}, \code{time_start}, \code{time_end}, \code{a_id}, \code{ar}).
+a dataframe with columns:  \code{tier}, \code{id}, \code{content}, \code{tier_name}, \code{tier_type}, \code{time_start}, \code{time_end}, \code{source}).
 }
 \description{
 Convert .eaf file from ELAN to a dataframe.
