### 2020-06-28 1503d08
correct draw_sound and docs
