### 2020-06-29 b983421
make it possible to visualise all sound annotation files with the draw_sound() function
