+++ man/draw_sound.Rd
@@ -83,6 +83,9 @@ Create oscilogram and spectrogram plot.
 \examples{
 draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"))
 
+draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+           system.file("extdata", "test.TextGrid", package = "phonfieldwork"))
+
 }
 \author{
 George Moroz <agricolamz@gmail.com>
