### 2020-06-30 fa56313
add bridge to lingtypology package: "map" argument in the "create_viewer()" function; close #17
