+++ man/create_viewer.Rd
@@ -11,6 +11,7 @@ create_viewer(
   captions = NULL,
   sorting_columns = NULL,
   about = "Created with the `phonfieldworks` package (Moroz 2020).",
+  map = FALSE,
   output_dir,
   output_file = "stimuli_viewer",
   render = TRUE
@@ -28,6 +29,8 @@ create_viewer(
 \item{sorting_columns}{vector of strings for sorting the result column}
 
 \item{about}{it is either .Rmd file or string with the text for about information: author, project, place of gahtered information and other metadata, version of the viewer and so on}
+
+\item{map}{the logical argument, if \code{TRUE} and there is a \code{glottocode} column in \code{table}}
 
 \item{output_dir}{the output directory for the rendered file}
 
