+++ docs/bibliography.bib
@@ -6,6 +6,29 @@
   year={2003}
 }
 
+@Manual{reidy16,
+    title = {textgRid: Praat TextGrid Objects in R},
+    author = {Patrick Reidy},
+    year = {2016},
+    note = {R package version 1.0.1},
+    url = {https://CRAN.R-project.org/package=textgRid},
+  }
+
+@InProceedings{boril16,
+    author = {Tomáš Bořil and Radek Skarnitzl},
+    editor = {Petr Sojka and Aleš Horák and Ivan Kopeček and Karel Pala},
+    booktitle = {Text, Speech, and Dialogue: 19th International Conference, TSD 2016, Brno, Czech Republic, September 12-16, 2016, Proceedings},
+    publisher = {Springer International Publishing},
+    address = {Cham},
+    year = {2016},
+    title = {Tools rPraat and mPraat},
+    pages = {367--374},
+    isbn = {978-3-319-45510-5},
+    doi = {10.1007/978-3-319-45510-5_42},
+    url = {http://dx.doi.org/10.1007/978-3-319-45510-5_42},
+  }
+
+
 @misc{lubbers13,
 author={Lubbers, Mart and Torreira, Francisco},
 year={2013},
@@ -18,4 +41,30 @@ and {P}raats {T}ext{G}rid annotation files.}
   author={Bowern, Claire},
   year={2015},
   publisher={Springer}
+}
+
+@Manual{moroz17,
+    title = {lingtypology: easy mapping for Linguistic Typology},
+    author = {George Moroz},
+    year = {2017},
+    url = {https://CRAN.R-project.org/package=lingtypology},
+  }
+
+@article{schmidt09,
+  title={EXMARaLDA--Creating, analysing and sharing spoken language corpora for pragmatic research},
+  author={Schmidt, Thomas and W{\"o}rner, Kai},
+  journal={Pragmatics},
+  volume={19},
+  number={4},
+  pages={565--582},
+  year={2009},
+  publisher={John Benjamins}
+}
+
+@inproceedings{wittenburg06,
+  title={ELAN: a professional framework for multimodality research},
+  author={Wittenburg, Peter and Brugman, Hennie and Russel, Albert and Klassmann, Alex and Sloetjes, Han},
+  booktitle={5th International Conference on Language Resources and Evaluation (LREC 2006)},
+  pages={1556--1559},
+  year={2006}
 }
