+++ docs/bibliography.bib
@@ -68,3 +68,11 @@ and {P}raats {T}ext{G}rid annotation files.}
   pages={1556--1559},
   year={2006}
 }
+
+
+@misc{hammarstrom2020,
+  title={Glottolog 4.2},
+  author={Hammarstr{\"o}m, Harald and Forkel, Robert and Haspelmath, Martin},
+  note={Jena: Max Planck Institute for the Science of Human History},
+  year={2020}
+}
