### 2020-07-01 9ee9f0d
add possibility to map by lat-long
