### 2020-07-01 agricolamz/phonfieldwork#19 @agricolamz
When mapping click in the table doesn't work
Correct multiblock problem.
