### 2020-07-02 a7b13b8
add window annotation to spectrograms
