+++ R/draw_spectrogram.R
@@ -12,21 +12,24 @@
 #' @param padding The amount of zero padding for each window, measured in units of window length. For example, if the window is 50 points, and padding = 10, 500 zeros will be appended to each window.
 #' @param preemphasisf Preemphasis of 6 dB per octave is added to frequencies above the specified frequency. For no preemphasis, set to a frequency higher than the sampling frequency.
 #' @param maxfreq the maximum frequency to be displayed for the spectrogram up to a maximum of fs/2. This is set to 5000 Hz by default.
-#' @param colors If TRUE, a color spectrogram will be displayed. If FALSE, greyscale is used. If a vector of colors is provided, these colors are used to create the spectrogram.
 #' @param dynamicrange Values greater than this many dB below the maximum will be displayed in the same color.
 #' @param nlevels The number of divisions to be used for the z-axis of the spectrogram. By default it is set equal to the dynamic range, meaning that a single color represents 1 dB on the z-axis.
 #' @param show If FALSE, no spectrogram is plotted. This is useful if the user would like to perform an action on an existing spectrogram plot without having to redraw it.
-#' @param window the window to be applied to the signal, applied by the windowfunc function in this package.
-#' @param windowparameter the parameter for the window to be applied to the signal, if appropriate.
-#' @param quality If TRUE, a contour plot is created, which results in a high-quality image that may be slow to plot. If FALSE, a lower-quality image is created that plots much faster.
+#' @param window A string indicating the type of window desired. Supported types are: rectangular, hann, hamming, cosine, bartlett, gaussian, and kaiser.
+#' @param windowparameter The parameter necessary to generate the window, if appropriate. At the moment, the only windows that require parameters are the Kaiser and Gaussian windows. By default, these are set to 2 for kaiser and 0.4 for gaussian windows.
 #' @param x_axis If TRUE then draw x axis.
 #' @param title Character with the title.
+#'
+#' @examples
+#' draw_spectrogram(system.file("extdata", "test.wav", package = "phonfieldwork"))
 #'
 #' @export
 #'
 #' @importFrom stats fft
 #' @importFrom tuneR readWave
 #' @importFrom tuneR readMP3
+#' @importFrom grDevices colorRampPalette
+#' @importFrom graphics image
 #'
 
 draw_spectrogram <- function (sound,
@@ -37,19 +40,18 @@ draw_spectrogram <- function (sound,
                               padding = 10,
                               preemphasisf = 50,
                               maxfreq = 5000,
-                              colors = FALSE,
                               dynamicrange = 50,
                               nlevels = dynamicrange,
                               show = TRUE,
                               window = "kaiser",
-                              windowparameter = 3,
-                              quality = FALSE,
+                              windowparameter = -1,
                               x_axis = TRUE,
                               title = NULL){
 
   # This function is slightly modification of phonTools::spectrogram() by Santiago Barreda <sbarreda@ucdavis.edu>
   if(class(sound) != "integer"){
-    ext <- tolower(substring(sound, regexpr("\\..*$", sound) + 1))
+    ext <- unlist(strsplit(normalizePath(sound), "\\."))
+    ext <- ext[length(ext)]
 
[…]
