+++ NAMESPACE
@@ -25,11 +25,13 @@ export(set_textgrid_names)
 export(srt_to_df)
 export(textgrid_to_df)
 export(tier_to_df)
+importFrom(grDevices,colorRampPalette)
 importFrom(grDevices,dev.off)
 importFrom(grDevices,png)
 importFrom(grDevices,rgb)
 importFrom(graphics,abline)
 importFrom(graphics,axis)
+importFrom(graphics,image)
 importFrom(graphics,par)
 importFrom(graphics,points)
 importFrom(graphics,rect)

[…]
