### 2020-07-02 5d02ffc
add window annotation to spectrograms
