### 2020-07-03 d0c0dd2
replace example_textgrid with systemfile() call; close #20; add ability to remove spectrum info; add preemphasis info
