+++ R/df_to_tier.R
@@ -14,7 +14,9 @@
 #'                     time_start = c(0.00000000,0.01246583,0.24781914,0.39552363,0.51157715),
 #'                     time_end = c(0.01246583,0.24781914,0.39552363,0.51157715,0.65267574),
 #'                     content = c("", "T", "E", "S", "T"))
-#' df_to_tier(my_df, textgrid = example_textgrid, overwrite = FALSE)
+#' df_to_tier(my_df,
+#' system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
+#' overwrite = FALSE)
 #'
 #' @export
 #'
