+++ R/example_textgrid.R
@@ -1,8 +1,0 @@
-#' Praat TextGrid example file
-#'
-#' Praat TextGrid example file that has two tiers: an interval tier with 5 intervals and a point tier with 4 points.
-#'
-#' @format A character vector of the length 52
-#'
-
-"example_textgrid"
