+++ R/get_textgrid_names.R
@@ -9,7 +9,7 @@
 #'
 #' @return return a vector of tier names from given TextGrid
 #' @examples
-#' get_textgrid_names(example_textgrid)
+#' get_textgrid_names(system.file("extdata", "test.TextGrid", package = "phonfieldwork"))
 #'
 #' @export
 #'
