### 2020-07-03 90ceeb6
correct annotation problem; replace freqmax with frequency_range argument
