+++ docs/index.Rmd
@@ -313,7 +313,7 @@ There are additional parameters:
 * `zoom` -- time in seconds for zooming spectrogram
 * `text_size` -- size of the text on the plot
 * `annotation` -- the optional file with the TextGrid's file path or dataframe with annotations (see the section 5.)
-* `maximum_frequency` -- the maximum frequency to be displayed for the spectrogram
+* `frequency_range` -- the frequency range to be displayed for the spectrogram
 * `dynamic_range` -- values greater than this many dB below the maximum will be displayed in the same color
 * `window_length` -- the desired length in milliseconds for the analysis window 
 * `window` -- window type (can be "rectangular", "hann", "hamming", "cosine", "bartlett", "gaussian", and "kaiser")
