+++ R/draw_spectrogram.R
@@ -13,13 +13,14 @@
 #' @param timestep If a negative value is given, -N, then N equally-spaced time steps are calculated. If a positive number is given, this is the spacing between adjacent analyses, in milliseconds.
 #' @param padding The amount of zero padding for each window, measured in units of window length. For example, if the window is 50 points, and padding = 10, 500 zeros will be appended to each window.
 #' @param preemphasisf Preemphasis of 6 dB per octave is added to frequencies above the specified frequency. For no preemphasis, set to a frequency higher than the sampling frequency.
-#' @param frequency_range vector with the range of frequencies to be displayed for the spectrogram up to a maximum of fs/2. This is set to 0-5 kHz by default.
+#' @param frequency_range vector with the range of frequencies to be displayed for the spectrogram up to a maximum of \code{fs}/2. This is set to 0-5 kHz by default.
 #' @param dynamic_range Values greater than this many dB below the maximum will be displayed in the same color.
 #' @param nlevels The number of divisions to be used for the z-axis of the spectrogram. By default it is set equal to the dynamic range, meaning that a single color represents 1 dB on the z-axis.
 #' @param window A string indicating the type of window desired. Supported types are: rectangular, hann, hamming, cosine, bartlett, gaussian, and kaiser.
 #' @param windowparameter The parameter necessary to generate the window, if appropriate. At the moment, the only windows that require parameters are the Kaiser and Gaussian windows. By default, these are set to 2 for kaiser and 0.4 for gaussian windows.
 #' @param x_axis If \code{TRUE} then draw x axis.
 #' @param title Character with the title.
+#' @param raven_annotation Raven (Center for Conservation Bioacoustics) style annotations (boxes over spectrogram). The dataframe that contains \code{time_start}, \code{time_end}, \code{freq_low} and \code{freq_high} columns. Optional columns are \code{colors} and \code{content}.
 #'
 #' @examples
 #' draw_spectrogram(system.file("extdata", "test.wav", package = "phonfieldwork"))
@@ -31,6 +32,7 @@
 #' @importFrom tuneR readMP3
 #' @importFrom grDevices colorRampPalette
 #' @importFrom graphics image
+#' @importFrom graphics rect
 #'
 
 draw_spectrogram <- function (sound,
@@ -48,7 +50,8 @@ draw_spectrogram <- function (sound,
                               frequency_range = c(0, 5),
                               nlevels = dynamic_range,
                               x_axis = TRUE,
-                              title = NULL){
+                              title = NULL,
+                              raven_annotation = NULL){
 
   # This function is slightly modification of phonTools::spectrogram() by Santiago Barreda <sbarreda@ucdavis.edu>
   if(class(sound) != "integer"){
@@ -167,5 +170,33 @@ draw_spectrogram <- function (sound,
   if(x_axis){
     graphics::axis(1, cex.axis=text_size, las=1)
     graphics::title(xlab = "Time (s)")
+  }
+  if(!is.null(raven_annotation)){
+    if(is.null(raven_annotation$colors)){
+      raven_annotation$colors <- "black"
+    }
+    if("time_start" %in% names(raven_annotation) &
+       "time_end" %in% names(raven_annotation) &
+       "freq_low" %in% names(raven_annotation) &
+       "freq_high" %in% names(raven_annotation)){
+      graphics::rect(xleft = raven_annotation$time_start,
+                     xright = raven_annotation$time_end,
+                     ybottom = raven_annotation$freq_low,
+                     ytop = raven_annotation$freq_high,
+                     lwd = 2,
+                     border = raven_annotation$colors)
+      if("content" %in% names(raven_annotation)){
+        graphics::text(x = raven_annotation$time_start,
+                       y = raven_annotation$freq_high,
[…]
