+++ docs/index.Rmd
@@ -365,6 +365,32 @@ tree | tail -n 25 | head -n 23
 
 It is also possible to use the argument `textgrid_from_folder` in order to specify the folder where .TextGrids for annotation are (could be the same folder as the sound one). By default the `draw_sound()` function with the `sounds_from_folder` argument adds a title with the file name to each pictures' title, but it is possible to turn it off using the argument `title_as_filename = FALSE`.
 
+If you are familiar with the Raven program for bioacoustics, you probably miss an ability to annotate not only time, but also a frequency range. In order to do it you need to create a dataframe with the columns `time_start`, `time_end`, `freq_low` and `freq_high`:
+
+```{r}
+raven_an <- data.frame(time_start = 450, 
+                       time_end  = 520, 
+                       freq_low = 3, 
+                       freq_high = 4.5)
+
+draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+           raven_annotation = raven_an)
+```
+
+It is also possible to use multiple values, colors (adding `colors` column) and annotation (adding `content` column):
+
+```{r}
+raven_an <- data.frame(time_start = c(250, 450), 
+                       time_end  = c(400, 520), 
+                       freq_low = c(1, 3), 
+                       freq_high = c(2, 4.5),
+                       colors = c("red", "blue"),
+                       content = c("a", "b"))
+
+draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+           raven_annotation = raven_an)
+```
+
 # Read linguistic files into R
 
 The `phonfieldwork` package provides also several methods for reading different file types into R. This makes it possible to analyze them and convert into `.csv` files (e. g. using the `write.csv()` function). The main advantage of using those functions is that all of them return `data.frame`s with columns (`time_start`, `time_end`, `content` and `source`). This make it easer to use the result in the `draw_sound()` function that make it possible to visualise all kind of sound annotation systems.
