### 2020-07-06 #385 @elinw
I think this is an interesting question.

Essentially this is a tool.  Could the tool be used by someone who has not adequately protected their human participants from being identified or who has not obtained voluntary informed consent? Yes, however I don't think there is anything that can reasonably be done about that by the tool maker because that would really require (from my understanding) changes to the data collection process.  Complicating this is the issue that while US-based researchers at federally funded institutions are required to comply with the Common Rule as well as the Belmont Report principles it was derived from, researchers in other places are subject to their local regulations.  And non-researchers have no such requirements at all.

The Common Rule and the Belmont Report do not create bright lines, they talk about balancing competing principles: beneficence, justice and respect for persons are sometimes not in agreement.  The harm must be weighed against the benefit. In the case of this kind of linguistic data, as I understand it, the potential harm is very much going to be context-dependent. First there is the issue of participants being identifiable, but, second, there could also be larger political and social issues that we have not thought about (for example when a language or culture is being suppressed).  This is why respect for persons requires that all of the potential risks be explained to potential participants before data collection.

Overall, I do think that having some kind of notice about the need to obtain voluntary informed consent will at least foreground the issue for researchers.  There could be a link to a larger document that describes the range of possible issues in more depth.
