### 2020-07-06 3c4bc8e
correct eaf reading; add argument eafs_from_folder
