### 2020-07-06 #385 @agricolamz
I'm totally agree, but that means also that the same kind of note should be put on each package that works with sounds. And for now our concern is only about humans, but we could think on it in a greater scale: imagine that I recorded and published audio with rhinoceros' mating signal that will make it easier for poachers to lure those endangered creatures...
