### 2020-07-07 #385 @annakrystalli
Thank you for your feedback @elinw. And thank you for your patience and response @agricolamz. As we encounter new applications involving human subject data, it might take a little longer and input from domain experts like @elinw might be required to help ensure our policy is fit for purpose.

So I'll summarise where I think we are at the moment:

Overall I think just some more detailed documentation regarding data protection for privacy + the handling data according to the informed consent will be enough to start the review process. In particular drawing attention to what @agricolamz highlighted:

> I think that there are several things that participant should agree on:
>
> - conduct scientific research on gathered data;
> - publication of gathered data;

I think you also made some good suggestions regarding encryption. Other simpler options could be warnings, asking for confirmation before performing an action (e.g. publishing data), adding particular files to .gitignore or even scanning data for potentially sensitive information (eg names).

I think a full review seems a better way to assess what strategy would be more useful and where throughout a typical workflow. So I suggest we ask reviewers to consider these points during review?

So, once some more documentation is added to the README on:

- [ ] A short demo of package functionality (ideally the README should also contain a short demo so potential users can skim functionality in the repo, then go to the longer vignette (tutorial) for more details.
- [ ] A summary of how the package compares to other packages and why it is best in class
- [ ] Information and guidance on data privacy and informed consent (it's fine to just summarise the main points and link to more detailed information)

I am happy to pass it on to the handling editor.
