+++ DESCRIPTION
@@ -22,3 +22,6 @@ BugReports: https://github.com/agricolamz/phonfieldwork/issues
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
+VignetteBuilder: knitr
+Suggests: 
+    knitr
