### 2020-07-08 b0c7431
add message about ethics
