+++ R/create_viewer.R
@@ -27,7 +27,7 @@ create_viewer <- function(audio_dir,
                           table,
                           captions = NULL,
                           sorting_columns = NULL,
-                          about = "Created with the `phonfieldworks` package (Moroz 2020).",
+                          about = "Created with the `phonfieldworks` package (Moroz 2020). <br> If you are a participant of this study and you want to withdraw your consent, please, write to ... on the following adress ...",
                           map = FALSE,
                           output_dir,
                           output_file = "stimuli_viewer",
