### 2020-07-10 3988acf
correct README for rOpenSci
