### 2020-07-10 #385 @agricolamz
@annakrystalli, @elinw,

I've added a message to the `.html` creating function that links to the [vigniette](https://github.com/agricolamz/phonfieldwork/blob/master/vignettes/ethical_research_with_phonfieldwork.Rmd), that I've created in order to emphasize an ethical issues. It would be nice if you take a look.

Since my package is actually a bunch of small functions I don't know how to make a small demo of it, but there is a link to the [whole tutorial](https://agricolamz.github.io/phonfieldwork/).

I also added a small summary and links to other packages.
