+++ R/exb_to_df.R
@@ -5,6 +5,7 @@
 #' @author George Moroz <agricolamz@gmail.com>
 #'
 #' @param file_name string with a filename or path to the .exb file
+#' @param exbs_from_folder path to a folder with multiple .exb files. If this argument is not \code{NULL}, then the function goes through all files and create a merged dataframe for all of them.
 #' @return a dataframe with columns:  \code{tier}, \code{id}, \code{content}, \code{tier_name}, \code{tier_type}, \code{tier_category}, \code{tier_speaker}, \code{time_start}, \code{time_end}, \code{source}.
 #'
 #' @examples
@@ -17,62 +18,69 @@
 #' @importFrom xml2 xml_text
 #' @importFrom xml2 xml_children
 
-exb_to_df <- function(file_name){
-  # read file
-  l <- xml2::read_xml(file_name)
-  # extract tiers
-  t <- xml2::xml_find_all(l, "basic-body/tier")
-  # tier names
-  tier_names <- xml2::xml_attr(t, "display-name")
-  # tier types
-  tier_types <- xml2::xml_attr(t, "type")
-  # tier category
-  tier_categories <- xml2::xml_attr(t, "category")
-  # tier category
-  tier_speakers <- xml2::xml_attr(t, "speaker")
+exb_to_df <- function(file_name, exbs_from_folder = NULL){
+  if(is.null(exbs_from_folder)){
+    # read file
+    l <- xml2::read_xml(file_name)
+    # extract tiers
+    t <- xml2::xml_find_all(l, "basic-body/tier")
+    # tier names
+    tier_names <- xml2::xml_attr(t, "display-name")
+    # tier types
+    tier_types <- xml2::xml_attr(t, "type")
+    # tier category
+    tier_categories <- xml2::xml_attr(t, "category")
+    # tier category
+    tier_speakers <- xml2::xml_attr(t, "speaker")
 
-  # create list of dataframes
-  lapply(seq_along(t), function(i){
-    content <- xml2::xml_text(xml2::xml_find_all(t[[i]], "event"))
-    ts1 <- xml2::xml_attr(xml2::xml_children(t[[i]]), "start")
-    ts2 <- xml2::xml_attr(xml2::xml_children(t[[i]]), "end")
-    data.frame(tier = i,
-               id = 1:length(content),
-               content = content,
-               tier_name = tier_names[i],
-               tier_type = tier_types[i],
-               tier_category = tier_categories[i],
-               tier_speaker = tier_speakers[i],
-               ts_start = ts1,
-               ts_end = ts2,
-               stringsAsFactors = FALSE)
-  }) ->
-    r
-  # merge list of dataframes  into dataframe
-  r <- do.call(rbind, r)
+    # create list of dataframes
+    lapply(seq_along(t), function(i){
+      content <- xml2::xml_text(xml2::xml_find_all(t[[i]], "event"))
+      ts1 <- xml2::xml_attr(xml2::xml_children(t[[i]]), "start")
+      ts2 <- xml2::xml_attr(xml2::xml_children(t[[i]]), "end")
+      data.frame(tier = i,
+                 id = 1:length(content),
+                 content = content,
+                 tier_name = tier_names[i],
+                 tier_type = tier_types[i],
+                 tier_category = tier_categories[i],
+                 tier_speaker = tier_speakers[i],
+                 ts_start = ts1,
+                 ts_end = ts2,
+                 stringsAsFactors = FALSE)
+    }) ->
+      r
+    # merge list of dataframes  into dataframe
+    r <- do.call(rbind, r)
 
-  # extract info about time
[…]
