+++ NAMESPACE
@@ -33,6 +33,7 @@ importFrom(graphics,abline)
 importFrom(graphics,axis)
 importFrom(graphics,image)
 importFrom(graphics,par)
+importFrom(graphics,plot)
 importFrom(graphics,points)
 importFrom(graphics,rect)
 importFrom(graphics,segments)

[…]
