+++ DESCRIPTION
@@ -23,5 +23,4 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.0
 VignetteBuilder: knitr
-Suggests: 
-    knitr
+Suggests: knitr
