+++ travis.yml
@@ -1,0 +1,6 @@
+language: r
+cache: packages
+notifications:
+  email:
+    on_success: never
+    on_failure: always
