+++ README.md
@@ -3,6 +3,7 @@
 [![Project Status: Active - The project has reached a stable, usable state and is being actively developed.](http://www.repostatus.org/badges/latest/active.svg)](http://www.repostatus.org/#active)
 [![CRAN version](http://www.r-pkg.org/badges/version/phonfieldwork)](https://cran.r-project.org/package=phonfieldwork)
 [![](http://cranlogs.r-pkg.org/badges/grand-total/phonfieldwork)](https://CRAN.R-project.org/package=phonfieldwork)
+[![](https://travis-ci.org/agricolamz/phonfieldwork.svg?branch=master)](https://travis-ci.org/github/agricolamz/phonfieldwork)
 [![DOI](https://zenodo.org/badge/194053227.svg)](https://zenodo.org/badge/latestdoi/194053227)
 
 `phonfieldwork` is a package for phonetic fieldwork research and experiments. This package make it easier:
