### 2020-07-11 ce41630
add pitch arguments to the draw_sound() function
