+++ R/draw_sound.R
@@ -40,6 +40,11 @@
 #'
 #' draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
 #'            system.file("extdata", "test.TextGrid", package = "phonfieldwork"))
+#'
+#' draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+#'            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
+#'            pitch = system.file("extdata", "test.Pitch", package = "phonfieldwork"),
+#'            pitch_range = c(50, 200))
 #' }
 #' @export
 #'
@@ -77,6 +82,8 @@ draw_sound <- function(file_name,
                        preemphasisf = 50,
                        spectrum_info = TRUE,
                        raven_annotation = NULL,
+                       pitch = NULL,
+                       pitch_range = c(75, 350),
                        output_width = 750,
                        output_height = 500,
                        output_units = "px",
@@ -136,7 +143,7 @@ draw_sound <- function(file_name,
       title_space <- ifelse(is.null(title), 0, 2)
 
       # plot oscilogram ---------------------------------------------------------
-      low_boundary <- ifelse(is.null(zoom), 0.75, 0.83)
+      low_boundary <- ifelse(is.null(zoom), 0.83, 0.91)
 
       graphics::par(oma=c(0, 0,title_space,0),
                     mai=c(0, 0.8, 0, 0.2),
@@ -164,8 +171,18 @@ draw_sound <- function(file_name,
         graphics::axis(1, las=1)
       }
       # plot spectrogram --------------------------------------------------------
-      low_boundary <- ifelse(is.null(annotation), 0.1, 0.27)
-      graphics::par(fig=c(0, 0.97, low_boundary, 0.75), new=TRUE)
+
+      if(is.null(annotation) & is.null(pitch)){
+        low_boundary <- 0.1
+      } else if(is.null(annotation) & !is.null(pitch)){
+        low_boundary <- 0.1 + 0.17
+      } else if(!is.null(annotation) & is.null(pitch)){
+        low_boundary <- 0.1 + 0.17
+      } else if(!is.null(annotation) & !is.null(pitch)){
+        low_boundary <- 0.1 + 0.17 + 0.17
+      }
+
+      graphics::par(fig=c(0, 0.97, low_boundary, 0.83), new=TRUE)
       if(!is.null(zoom)){
         for_spectrum <- tuneR::extractWave(s,
                                            from = zoom[1],
@@ -186,11 +203,57 @@ draw_sound <- function(file_name,
                        preemphasisf = preemphasisf,
                        spectrum_info = spectrum_info,
                        raven_annotation = raven_annotation,
-                       x_axis = is.null(annotation))
+                       x_axis = low_boundary == 0.1)
+
+
+      # plot pitch --------------------------------------------------------------
+      if(is.null(annotation) & !is.null(pitch)){
+        upper_boundary <- 0.1 + 0.17
+        low_boundary <- 0.1
+      } else if(!is.null(annotation) & is.null(pitch)){
+        upper_boundary <- 0.1 + 0.17
+        low_boundary <- 0.1
+      } else if(!is.null(annotation) & !is.null(pitch)){
+        upper_boundary <- 0.1 + 0.17 + 0.17
+        low_boundary <- 0.1+ 0.17
+      }
+
+      if(!is.null(pitch)){
+        graphics::par(fig=c(0, 0.97, low_boundary, upper_boundary), new=TRUE)
+        if(class(pitch) != "data.frame"){
+          pitch <- phonfieldwork::pitch_to_df(pitch)
+        }
+
+        graphics::plot(pitch$time_start*1000, pitch$frequency,
[…]
