+++ man/pitch_to_df.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/pitch_to_list.R
+% Please edit documentation in R/pitch_to_df.R
 \name{pitch_to_df}
 \alias{pitch_to_df}
 \title{Praat Pitch tier to dataframe}
@@ -18,6 +18,10 @@ a dataframe with columns:  \code{time_start}, \code{time_end}, \code{frequency} 
 }
 \description{
 Convert a Praat Pitch tier to a dataframe.
+}
+\examples{
+pitch_to_df(system.file("extdata", "test.Pitch", package = "phonfieldwork"))
+
 }
 \author{
 George Moroz <agricolamz@gmail.com>
