+++ vignettes/introduction_to_phonfieldwork.Rmd
@@ -1,20 +1,16 @@
 ---
-title: "Phonetic fieldwork and experiments with phonfieldwork package"
+title: "Phonetic fieldwork and experiments with `phonfieldwork` package"
 author: "G. Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
 bibliography: bibliography.bib
-date: "`r Sys.Date()`"
+date: "2020-07-11"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Phonetic fieldwork and experiments with phonfieldwork package}
+  %\VignetteIndexEntry{Phonetic fieldwork and experiments with `phonfieldwork` package}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
 
-```{r setup, include=FALSE}
-knitr::opts_chunk$set(eval = FALSE)
-```
 
-The last and more detailed version of this file can be found [here](https://agricolamz.github.io/phonfieldwork/).
 
 # Introduction
 
@@ -26,13 +22,15 @@ There are a lot of different books about linguistic fieldwork and experiments (e
 
 Before you start, make sure that you have installed the package, for example with the following command:
 
-```{r}
+
+```r
 install.packages("phonfieldwork")
 ```
 
 This command will install the last stable version of the `phonfieldwork` package from CRAN. Since CRAN runs multiple package checks before making it available, this is the safest option. Alternatively, you can download the development version from GitHub:
 
-```{r}
+
+```r
 install.packages("devtools")
 devtools::install_github("agricolamz/phonfieldwork")
 ```
@@ -40,19 +38,41 @@ devtools::install_github("agricolamz/phonfieldwork")
 If you have any trouble installing the package, you will not be able to use its functionality. In that case you can create [an issue on Github](https://github.com/agricolamz/phonfieldwork/issues) or send an email. Since this package could completely destroy your data, **please do not use it until you are sure that you have made a backup**.
 
 Use the `library()` command to load the package:
-```{r, eval=TRUE}
+
+```r
 library("phonfieldwork")
 ```
 
 This tutorial was made using the following version of `phonfieldwork`:
-```{r, eval=TRUE}
+
+```r
 packageVersion("phonfieldwork")
+```
+
+```
+## [1] '0.0.8'
 ```
 
 This tutorial can be cited as follows:
 
-```{r, eval=TRUE}
+
+```r
 citation("phonfieldwork")
+```
+
+```
+## 
+## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork package_. <URL:
+## https://CRAN.R-project.org/package=phonfieldwork>.
+## 
+## A BibTeX entry for LaTeX users is
+## 
+##   @Manual{,
+##     title = {Phonetic fieldwork and experiments with phonfieldwork package},
+##     author = {George Moroz},
+##     year = {2020},
+##     url = {https://CRAN.R-project.org/package=phonfieldwork},
+##   }
 ```
 
 If you have any trouble using the package, do not hesitate to create [an issue on Github](https://github.com/agricolamz/phonfieldwork/issues/new).
@@ -73,7 +93,7 @@ Most phonetic research consists of the following steps:
 
 The `phonfieldwork` package is created for helping with items 3, partially with 4, and 5 and 8.
 
[…]
