+++ vignettes/phonfieldwork.Rmd
@@ -162,7 +162,7 @@ create_presentation(stimuli = my_stimuli_df$stimuli,
 
 As a result, a file "first_example.html" was created in the output folder. You can change the name of this file by changing  the `output_file` argument. The `.html` file now looks as follows:
 
-<iframe src="https://agricolamz.github.io/phonfieldwork/first_example.html" width = 800 height = 600>
+<iframe src="https://agricolamz.github.io/phonfieldwork/additional/first_example.html" width = 800 height = 600>
   <p>Your browser does not support iframes :(</p>
 </iframe>
 
@@ -173,14 +173,14 @@ After collecting data and removing soundfiles with unsuccesful elicitations, one
 
 
 ```
-## ├── s1
-## │   ├── 01.wav
-## │   ├── 02.wav
-## │   └── 03.wav
-## └── s2
-##     ├── 01.wav
-##     ├── 02.wav
-##     └── 03.wav
+## ├── unnamed-chunk-26-1.png
+## ├── unnamed-chunk-28-1.png
+## ├── unnamed-chunk-32-1.png
+## ├── unnamed-chunk-33-1.png
+## ├── unnamed-chunk-34-1.png
+## ├── unnamed-chunk-39-1.png
+## ├── unnamed-chunk-40-1.png
+## └── unnamed-chunk-49-1.png
 ```
 
 For each speaker `s1` and `s2` there is a folder that containes three audiofiles. Now let's rename the files.
@@ -196,18 +196,18 @@ As a result, you obtain the following structure:
 
 
 ```
-## ├── s1
-## │   ├── backup
-## │   │   ├── 01.wav
-## │   │   ├── 02.wav
-## │   │   └── 03.wav
-## │   ├── s1_tap.wav
-## │   ├── s1_tip.wav
-## │   └── s1_top.wav
-## └── s2
-##     ├── 01.wav
-##     ├── 02.wav
-##     └── 03.wav
+## │   └── 03.wav
+## ├── unnamed-chunk-20-1.png
+## ├── unnamed-chunk-22-1.png
+## ├── unnamed-chunk-24-1.png
+## ├── unnamed-chunk-26-1.png
+## ├── unnamed-chunk-28-1.png
+## ├── unnamed-chunk-32-1.png
+## ├── unnamed-chunk-33-1.png
+## ├── unnamed-chunk-34-1.png
+## ├── unnamed-chunk-39-1.png
+## ├── unnamed-chunk-40-1.png
+## └── unnamed-chunk-49-1.png
 ```
 
 The `rename_soundfiles()` function created a backup folder with all of the unrenamed files, and renamed all files using the prefix provided in the `prefix` argument. There is an additional argument `backup` that can be set to `FALSE` (it is `TRUE` by default), in case you are sure that the renaming function will work properly with your files and stimuli, and you do not need a backup of the unrenamed files.
@@ -223,18 +223,18 @@ rename_soundfiles(stimuli = my_stimuli_df$stimuli,
 
 
 ```
-## ├── s1
-## │   ├── backup
-## │   │   ├── 01.wav
-## │   │   ├── 02.wav
-## │   │   └── 03.wav
-## │   ├── s1_tap.wav
-## │   ├── s1_tip.wav
-## │   └── s1_top.wav
-## └── s2
-##     ├── s2_tap_2.wav
-##     ├── s2_tip_1.wav
-##     └── s2_top_3.wav
+## │   └── s2_top_3.wav
+## ├── unnamed-chunk-20-1.png
+## ├── unnamed-chunk-22-1.png
[…]
