+++ _pkgdown.yml
@@ -1,1 +1,1 @@
-url: https://agricolamz.github.io/phonfieldwork
+url: https://agricolamz.github.io/phonfieldwork/index.html
