+++ _pkgdown.yml
@@ -1,1 +1,2 @@
-url: https://agricolamz.github.io/phonfieldwork/index.html
+destination: docs
+url: https://agricolamz.github.io/phonfieldwork/
