### 2020-07-11 #385 @agricolamz
@melvidoni, I decided to rebuild documentation with `pkgdown` but for some reason it [works locally, but not on githubpages](https://github.com/agricolamz/phonfieldwork/issues/22)... I'll fix it soon.
