### 2020-07-11 ef85f26
fix tree visualisation
