### 2020-07-12 76ee6e6
correct NEWS.md
