+++ NEWS.md
@@ -1,9 +1,9 @@
-# phonfieldwork 0.0.8 (11.07.2020)
+# phonfieldwork 0.0.8
 
 - add the "picth" and "pitch_range" arguments to the "draw_sound()" function
 - add the "picth_to_df()" function
 
-# phonfieldwork 0.0.7 (10.07.2020)
+# phonfieldwork 0.0.7
 
 - add a vigniettes about ethical research and introduction to work with phonfieldwork
 - add an argument "textgrids_from_folder" to the "textgrid_to_df()" function
@@ -21,7 +21,7 @@
 - chage textgrid related functions' output from "start", "end", "annotation" to "time_start", "time_end", "content"
 - correct point tier visualization
 
-# phonfieldwork 0.0.6 (20.06.2020)
+# phonfieldwork 0.0.6
 
 - add "encoding" arguments to functions for working with TextGrids
 - add "textgrid" argument to the "draw_sound()" function
@@ -35,7 +35,7 @@
 - add an argument "title_as_filename" to the "draw_sound()" function
 - change .TextGrid associated arguments of the "create_viewer()" function to "table" argument; as a result users now need to provide a table for the annotation viewer and not a .TextGrid
 
-# phonfieldwork 0.0.5 (07.06.2020)
+# phonfieldwork 0.0.5
 
 - add "textgrid_to_df()" function for reading Praat files
 - add "create_glossed_document()" function for converting .flextext files into a glossed document
@@ -44,18 +44,18 @@
 - add "exb_to_df()" function for reading EXMARaLDA files
 - rename "textgrid" argument into "annotation" argument in "concatenate_soundfiles()" function adding new possible values
 
-# phonfieldwork 0.0.4 (10.05.2020)
+# phonfieldwork 0.0.4
 
 - add "create_subannotation()" function
 
-# phonfieldwork 0.0.3 (07.01.2020)
+# phonfieldwork 0.0.3
 
 - vertically and horisontally center text in presentations created by "create_presentation()"; thx @Pandaklez #1
 - add the "font_size" argument to the "create_presentation()" function
 - add "rename_videofiles()" function
 - rebuild html viewer for sounds with JavaScript with the help of new functions "create_image_look_up()" and "create_sound_play()".
 
-# phonfieldwork 0.0.2 (23.09.2019)
+# phonfieldwork 0.0.2
 
 - make the "create_presentation()" function render silently
 - add a new function "draw_sound()" for creating spectrogram and oscilogram
@@ -64,6 +64,6 @@
 - add new functions  "get_textgrid_names()" and "set_textgrid_names()"
 - finish tutorial <https://agricolamz.github.io/phonfieldwork/>
 
-# phonfieldwork 0.0.1 (26.08.2019)
+# phonfieldwork 0.0.1
 
 - initial release
