+++ _pkgdown.yml
@@ -1,2 +1,5 @@
 destination: docs
 url: https://agricolamz.github.io/phonfieldwork/
+template:
+  params:
+    bootswatch: readable
