### 2020-07-13 1d5a6bf
small changes in vignettes
