### 2020-07-13 #385 @melvidoni
### Editor checks:

---

#### Editor comments

Hello @agricolamz thanks for submitting! I'm Dr Melina Vidoni, and I'll be your handling editor. Here are some basic checks for your package:

```
> spelling::spell_check_package()
DESCRIPTION does not contain 'Language' field. Defaulting to 'en-US'
```

And the output for `googpractice::gp`:

```
── GP phonfieldwork ───────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions,
    and all package code in general. 0% of code
    lines are covered by test cases.

    R/add_leading_symbols.R:17:NA
    R/add_leading_symbols.R:18:NA
    R/annotate_textgrid.R:34:NA
    R/annotate_textgrid.R:35:NA
    R/annotate_textgrid.R:37:NA
    ... and 1105 more lines

  ✖ use '<-' for assignment instead of
    '='. '<-' is the standard, and R users and
    developers are used it and it is easier to read
    your code for them if you use '<-'.

    R\draw_spectrogram.R:73:5
    R\draw_spectrogram.R:74:16
    R\draw_spectrogram.R:75:31
    R\draw_spectrogram.R:76:32
    R\draw_spectrogram.R:78:11
    ... and 27 more lines

  ✖ avoid long code lines, it is bad for
    readability. Also, many people prefer editor
    windows that are about 80 characters wide. Try
    make your lines shorter than 80 characters

    R\add_leading_symbols.R:3:1
    R\annotate_textgrid.R:3:1
    R\annotate_textgrid.R:9:1
    R\annotate_textgrid.R:15:1
    R\annotate_textgrid.R:19:1
    ... and 131 more lines

  ✖ avoid sapply(), it is not type safe.
    It might return a vector, or a list, depending
    on the input data. Consider using vapply()
    instead.

    R\draw_spectrogram.R:93:11

  ✖ avoid the library() and require()
    functions, they change the global search path.
    If you need to use other packages, import them.
    If you need to load them explicitly, then
    consider loadNamespace() instead, or as a last
    resort, declare them as 'Depends' dependencies.

    R\create_viewer.R:113:1

  ✖ avoid 1:length(...), 1:nrow(...),
    1:ncol(...), 1:NROW(...) and 1:NCOL(...)
    expressions. They are error prone and result 1:0
    if the expression on the right hand side is
    zero. Use seq_len() or seq_along() instead.

    R\add_leading_symbols.R:18:40
    R\create_subannotation.R:46:10
    R\create_subannotation.R:54:23
    R\df_to_tier.R:52:31
    R\df_to_tier.R:61:31
    ... and 7 more lines

  ✖ fix this R CMD check ERROR: Running
    examples in 'phonfieldwork-Ex.R' failed The
    error most likely occurred in: > ### Name:
    srt_to_df > ### Title: Subtitles .srt file to
    dataframe > ### Aliases: srt_to_df > > ### **
    Examples > > srt_to_df(system.file("extdata",
    "test.srt", package = "phonfieldwork")) Error in
    strsplit(result[, "times"], " --> ") :
    non-character argument Calls: srt_to_df -> cbind
    -> do.call -> strsplit Execution halted
  ✖ fix this R CMD check WARNING: LaTeX
    errors when creating PDF version. This typically
    indicates Rd problems.
  ✖ fix this R CMD check ERROR: Re-running
    with no redirection of stdout/stderr. Hmm ...
    looks like a package Error in texi2dvi(file =
    file, pdf = TRUE, clean = clean, quiet = quiet,
    : pdflatex is not available Error in
    texi2dvi(file = file, pdf = TRUE, clean = clean,
    quiet = quiet, : pdflatex is not available Error
    in running tools::texi2pdf() You may want to
    clean up by 'rm -Rf
    C:/Users/e95207/AppData/Local/Temp/RtmpCIt1xR/Rd2pdfcf4c23af23e0'
```

Can you please fix these and advice when it is done? I will find reviewers after that.

---

Reviewers: TBD
Due date: TBD
