### 2020-07-14 dec7208
first corrections
