+++ R/pitch_to_df.R
@@ -5,10 +5,13 @@
 #' @author George Moroz <agricolamz@gmail.com>
 #'
 #' @param file_name string with a filename or path to the Pitch tier
-#' @param candidates Praat Pitch tier contains multiple candidates for each time slice, use the value \code{"all"} if you want to get them all
+#' @param candidates Praat Pitch tier contains multiple candidates for each
+#' time slice, use the value \code{"all"} if you want to get them all
 #' @param encoding Pitch tier encoding. Import from \code{readLines()} function.
 #'
-#' @return a dataframe with columns:  \code{time_start}, \code{time_end}, \code{frequency} and, if \code{candidates} = \code{"all"}, \code{candidate_id} and \code{strength}
+#' @return a dataframe with columns:  \code{time_start}, \code{time_end},
+#' \code{frequency} and, if \code{candidates} = \code{"all"},
+#' \code{candidate_id} and \code{strength}
 #'
 #' @examples
 #' pitch_to_df(system.file("extdata", "test.Pitch", package = "phonfieldwork"))
@@ -31,7 +34,7 @@ pitch_to_df <- function(file_name,
   time_end <- as.numeric(gsub("[^0-9\\.]", "", pitch[grep("xmax =", pitch)]))
 
   # get all frames ----------------------------------------------------------
-  s <- split(1:length(pitch), cumsum(grepl("frame \\[\\d{1,}\\]", pitch)))
+  s <- split(seq_along(pitch), cumsum(grepl("frame \\[\\d{1,}\\]", pitch)))
   s <- s[-1]
   lapply(seq_along(s), function(i){
     data.frame(candidate_id = 1:sum(grepl("frequency =", pitch[s[[i]]])),
@@ -45,7 +48,8 @@ pitch_to_df <- function(file_name,
 
   # define frames with maximum strengh --------------------------------------
   if(candidates != "all"){
-    result <- data.frame(frequency = result[result$candidate_id == 1, "frequency"])
+    result <- data.frame(frequency = result[result$candidate_id == 1,
+                                            "frequency"])
   }
   result$frequency <- ifelse(result$frequency == 0, NA, result$frequency)
   result$time_start <- seq(time_start, time_end, length.out = nrow(result))
