+++ R/create_image_look_up.R
@@ -3,8 +3,10 @@
 #' @author George Moroz <agricolamz@gmail.com>
 #'
 #' @param img_src string or vector of strings with a image(s) path(s).
-#' @param img_caption string or vector of strings that will be displayed when image is clicked.
-#' @param text string o vector of strings that will be displayed as view link. By default it is eye emoji (&#x1f441;).
+#' @param img_caption string or vector of strings that will be displayed when
+#' image is clicked.
+#' @param text string o vector of strings that will be displayed as view link.
+#' By default it is eye emoji (&#x1f441;).
 #' @return a string or vector of strings
 #'
 #' @examples
