+++ R/rename_soundfiles.R
@@ -9,8 +9,10 @@
 #' @param translations character vector of translations (optonal)
 #' @param prefix character vector of length one containing prefix for file names
 #' @param suffix character vector of length one containing suffix for file names
-#' @param order numeric vector that define the order of stimuli. By default the order of the stimuli is taken.
-#' @param backup logical. If TRUE, function creates backup folder with all files. By default is TRUE.
+#' @param order numeric vector that define the order of stimuli. By default the
+#' order of the stimuli is taken.
+#' @param backup logical. If TRUE, function creates backup folder with all
+#' files. By default is TRUE.
 #' @return no output
 #' @export
 #'
