### 2020-07-15 a8486e9
add some ideas to ehtic vignette
