### 2020-07-15 8a4aa6f
add some ideas to ehtic vignette
