### 2020-07-16 dd2403a
correct for rOpenSci review
