+++ .travis.yml
@@ -4,3 +4,9 @@ notifications:
   email:
     on_success: never
     on_failure: always
+
+r_packages:
+  - covr
+
+after_success:
+- Rscript -e 'covr::codecov()'
