+++ DESCRIPTION
@@ -23,5 +23,8 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
 VignetteBuilder: knitr
-Suggests: knitr
+Suggests: 
+    knitr,
+    testthat,
+    covr
 Language: en-US
