+++ tests/testthat/test-tier-to-df.R
@@ -2,6 +2,11 @@ library(phonfieldwork)
 context("Tests for 'tier_to_df()' function")
 
 file_1 <- system.file("extdata", "test.TextGrid", package = "phonfieldwork")
-test_that("tier_to_df",
-          expect_equal(tier_to_df(file_1)[,1], 1:5))
+file_2 <- readLines(system.file("extdata", "test.TextGrid",
+                                package = "phonfieldwork"))
+test_that("tier_to_df", {
+          expect_equal(tier_to_df(file_1)[,1], 1:5)
+          expect_equal(tier_to_df(file_2)[,1], 1:5)
+          expect_equal(tier_to_df(file_1, tier = 3)[,1], 1:4) # point tier
+  })
 
