### 2020-07-17 98809c7
correct flex_to_df(), close #23; start adding latex output to create_glossed_document()
