+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,19 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+Please, put the results of these functions below, if your issue is related to the technical bugs. Thank you!
+
+<details> <summary> info about OS and package versions </summary>
+```
+sessionInfo()$R.version$platform
+sessionInfo()$R.version$version.string
+packageVersion("rmarkdown")
+packageVersion("phonfieldork")
+```
+</details>
