### 2020-07-17 cf0682e
Delete ISSUE_TEMPLATE.Md
