+++ ISSUE_TEMPLATE.Md
@@ -1,10 +1,0 @@
-Please, put the results of these functions below, if your issue is related to the technical bugs. Thank you!
-
-<details> <summary> info about OS and package versions </summary>
-```
-sessionInfo()$R.version$platform
-sessionInfo()$R.version$version.string
-packageVersion("rmarkdown")
-packageVersion("phonfieldork")
-```
-</details>
