### 2020-07-17 90d49db
correct flextext functions
