+++ R/create_glossed_document.R
@@ -31,34 +31,29 @@ create_glossed_document <- function(flextext = NULL,
                                     output_format = "docx",
                                     example_pkg = NULL){
   if(!("dplyr" %in% utils::installed.packages()[,"Package"])){
-    stop('For this function you need to install dplyr package with a command
-         install.packages("dplyr").')
+    stop(paste0('For this function you need to install dplyr package with a',
+                'command install.packages("dplyr").'))
   }
   if(!("tidyr" %in% utils::installed.packages()[,"Package"])){
-    stop('For this function you need to install tidyr package with a command
-         install.packages("tidyr").')
+    stop(paste0('For this function you need to install tidyr package with a',
+                'command install.packages("tidyr").'))
   }
   if(!(output_format %in% c("html", "docx"))){
     stop('The output_format can be only "html" or "docx"')
   }
 
-  if(output_format %in% c("html", "docx")){
-    tmp1 <- tempfile(fileext = ".csv")
-    utils::write.csv(flextext_to_df(flextext), tmp1, row.names = FALSE)
-    output_format2 <- ifelse(output_format == "docx", "word", output_format)
-    rmarkdown::render(paste0(.libPaths()[1],
+  tmp1 <- tempfile(fileext = ".csv")
+  utils::write.csv(flextext_to_df(flextext), tmp1, row.names = FALSE)
+  output_format2 <- ifelse(output_format == "docx", "word", output_format)
+  rmarkdown::render(paste0(.libPaths()[1],
 "/phonfieldwork/rmarkdown/templates/glossed_document/skeleton/skeleton.Rmd"),
-                      params = list(data = tmp1, rows = rows,
-                                    example_pkg = example_pkg),
-                      output_dir = output_dir,
-                      output_format = paste0(output_format2[1], "_document"),
-                      quiet = TRUE,
-                      output_file = output_file)
-    message(paste0("Output created: ", normalizePath(output_dir), "/",
-                   output_file, ".", output_format[1]))
-    suppress_message <- file.remove(tmp1)
-  } else if(output_format == "philex"){
-
-
-  }
+                    params = list(data = tmp1, rows = rows,
+                                  example_pkg = example_pkg),
+                    output_dir = output_dir,
+                    output_format = paste0(output_format2[1], "_document"),
+                    quiet = TRUE,
+                    output_file = output_file)
+  message(paste0("Output created: ", normalizePath(output_dir), "/",
+                 output_file, ".", output_format[1]))
+  suppress_message <- file.remove(tmp1)
 }
