+++ vignettes/phonfieldwork.Rmd
@@ -660,13 +660,13 @@ head(flextext_to_df("files/zilo_test.flextext"))
 ```
 
 ```
-##   s_id p_id    txt     cf hn     gls                   msa                free_trans
-## 1    1    1     б-     б-  1      an Inflects any category Жил-был (у Гъули?) петух.
-## 2    1    1    ик1    ик1  1    быть                    гл Жил-был (у Гъули?) петух.
-## 3    1    1     -о     -о  1     pst               гл:Past Жил-был (у Гъули?) петух.
-## 4    1    1     -й     -й  5 cvb(pf)    гл:Converb/Perfect Жил-был (у Гъули?) петух.
-## 5    1    1 =гъоди =гъоди  1    =rep                  част Жил-был (у Гъули?) петух.
-## 6    1    1     б-     б-  1      an Inflects any category Жил-был (у Гъули?) петух.
+##   p_id s_id w_id    txt     cf hn     gls                   msa                free_trans
+## 1    1    1    1     б-     б-  1      an Inflects any category Жил-был (у Гъули?) петух.
+## 2    1    1    1    ик1    ик1  1    быть                    гл Жил-был (у Гъули?) петух.
+## 3    1    1    1     -о     -о  1     pst               гл:Past Жил-был (у Гъули?) петух.
+## 4    1    1    1     -й     -й  5 cvb(pf)    гл:Converb/Perfect Жил-был (у Гъули?) петух.
+## 5    1    1    1 =гъоди =гъоди  1    =rep                  част Жил-был (у Гъули?) петух.
+## 6    1    1    2     б-     б-  1      an Inflects any category Жил-был (у Гъули?) петух.
 ##                            text_title                                morph
 ## 1 2017.04 Fairytale about the rooster d7f713db-e8cf-11d3-9764-00c04f186933
 ## 2 2017.04 Fairytale about the rooster d7f713e8-e8cf-11d3-9764-00c04f186933
