### 2020-07-17 d905ba9
correct flextext functions
