+++ inst/rmarkdown/templates/glossed_document/skeleton/skeleton.Rmd
@@ -17,108 +17,235 @@ library(dplyr, warn.conflicts = FALSE)
 library(tidyr, warn.conflicts = FALSE)
 
 if(is.null(example_pkg)){
-df %>%
-  group_by(text_title, p_id, s_id, free_trans, w_id) %>%
-  summarise(txt = paste0(txt, collapse = ""),
-            cf = paste0(cf, collapse = "-"),
-            hn = paste0(hn, collapse = "-"),
-            msa = paste0(msa, collapse = "-"),
-            gls = paste0(gls, collapse = "-")) %>% 
-  group_by(text_title, p_id, s_id, free_trans) %>%
-  summarise(txt = paste0(txt, collapse = "&#x09;"),
-            cf = paste0(cf, collapse = "&#x09;"),
-            hn = paste0(hn, collapse = "&#x09;"),
-            msa = paste0(msa, collapse = "&#x09;"),
-            gls = paste0(gls, collapse = "&#x09;")) %>% 
-  mutate(txt = paste0(s_id, "\\.&#x09;", txt, "\n\n&#x09;"),
-         cf = paste0(cf, "\n\n&#x09;"),
-         hn = paste0(hn, "\n\n&#x09;"),
-         msa = paste0(msa, "\n\n&#x09;"),
-         gls = paste0(gls, "\n\n&#x09;"),
-         free_trans = paste0(free_trans, "\n\n")) %>% 
-  select(text_title, p_id, s_id, txt, cf, hn, msa, gls, free_trans) %>% 
-  ungroup() %>% 
-  add_row(tibble(text_title = unique(.$text_title),
-                 p_id = 0,
-                 s_id = 0,
-                 txt = paste0("\n\n## ", unique(.$text_title), "\n\n"),
-                 cf = NA,
-                 hn = NA,
-                 msa = NA,
-                 gls = NA,
-                 free_trans = NA), .before = 1)  %>% 
-  arrange(text_title, p_id, s_id) %>% 
-  select(-c(all_rows[!(all_rows %in% params$rows)])) %>% 
-  pivot_longer(names_to = "type", values_to = "value", cols = c("txt", params$rows, "free_trans")) %>% 
-  na.omit() %>% 
-  ungroup() %>% 
-  select(value) %>% 
-  unlist() %>% 
-  unname() %>% 
-  cat(sep = "")
-  } else if(example_pkg == "philex"){
+  df %>%
+    group_by(text_title, p_id, s_id, free_trans, w_id) %>%
+    summarise(txt = paste0(txt, collapse = ""),
+              cf = paste0(cf, collapse = "-"),
+              hn = paste0(hn, collapse = "-"),
+              msa = paste0(msa, collapse = "-"),
+              gls = paste0(gls, collapse = "-")) %>% 
+    group_by(text_title, p_id, s_id, free_trans) %>%
+    summarise(txt = paste0(txt, collapse = "&#x09;"),
+              cf = paste0(cf, collapse = "&#x09;"),
+              hn = paste0(hn, collapse = "&#x09;"),
+              msa = paste0(msa, collapse = "&#x09;"),
+              gls = paste0(gls, collapse = "&#x09;")) %>% 
+    mutate(txt = paste0(s_id, "\\.&#x09;", txt, "\n\n&#x09;"),
+           cf = paste0(cf, "\n\n&#x09;"),
+           hn = paste0(hn, "\n\n&#x09;"),
+           msa = paste0(msa, "\n\n&#x09;"),
+           gls = paste0(gls, "\n\n&#x09;"),
+           free_trans = paste0(free_trans, "\n\n")) %>% 
+    select(text_title, p_id, s_id, txt, cf, hn, msa, gls, free_trans) %>% 
+    ungroup() %>% 
[…]
