### 2020-07-17 05545a1
add some tests; correct glossed_document
