### 2020-07-18 414fe18
add some tests; correct test names
