+++ tests/testthat/test-get-textgrid-names.R
@@ -4,7 +4,7 @@ context("Tests for 'get_textgrid_names()' function")
 file1 <- system.file("extdata", "test.TextGrid", package = "phonfieldwork")
 file2 <- readLines(file1)
 
-test_that("tier_to_df", {
+test_that("get_textgrid_names", {
   expect_equal(get_textgrid_names(file1),
                c("intervals", "empty_intervals", "points"))
   expect_equal(get_textgrid_names(file2),
