+++ tests/testthat/test-create-sound-play.R
@@ -1,7 +1,7 @@
 library(phonfieldwork)
 context("Tests for 'create_sound_play()' function")
 
-test_that("tier_to_df", {
+test_that("create_sound_play", {
   expect_equal(nchar(create_sound_play("a.wav")), 117)
 })
 
