### 2020-07-18 fe6d340
add some tests
