### 2020-07-18 906411c
add some tests
