+++ .travis.yml
@@ -7,6 +7,9 @@ notifications:
 
 r_packages:
   - covr
+  - dplyr
+  - tidyr
+  - DT
 
 after_success:
 - Rscript -e 'covr::codecov()'
