### 2020-07-18 22c8263
add some tests
