+++ tests/testthat/test-srt-to-df.R
@@ -1,0 +1,7 @@
+library(phonfieldwork)
+context("Tests for 'srt_to_df()' functions")
+t <- srt_to_df(system.file("extdata", "test.srt", package = "phonfieldwork"))
+
+test_that("srt_to_df", {
+  expect_true(nrow(t) == 4)
+})
