+++ tests/testthat/test-flextext-to-df.R
@@ -1,6 +1,6 @@
 library(phonfieldwork)
 context("Tests for 'flextext_to_df()' functions")
-file_name <- "../../vignettes/files/zilo_test.flextext"
+file_name <- "https://raw.githubusercontent.com/agricolamz/phonfieldwork/master/vignettes/files/zilo_test.flextext"
 df <- flextext_to_df(file_name)
 test_that("flextext_to_df", {
   expect_true(nrow(df) == 1170)
