+++ tests/testthat/test-create-presentation.R
@@ -8,3 +8,5 @@ create_presentation(stimuli = 1:3, output_dir = tmp, render = FALSE)
 test_that("create_presentation", {
   expect_true("stimuli_presentation.html" %in% list.files(tmp))
 })
+
+rm(tmp)
