+++ tests/testthat/test-create-viewer.R
@@ -18,9 +18,24 @@ create_viewer(audio_dir = paste0(tmp, "/audio"),
               picture_dir = paste0(tmp, "/pic"),
               table = data.frame(id = 1:2),
               output_dir = paste0(tmp, "/result"))
+
+create_viewer(audio_dir = paste0(tmp, "/audio"),
+              picture_dir = paste0(tmp, "/pic"),
+              table = data.frame(glottocode = c("stan1293", "russ1263")),
+              output_file = "with_a_map",
+              output_dir = paste0(tmp, "/result"),
+              map = TRUE)
+
 test_that("create_viewer", {
-  expect_true(length(list.files(paste0(tmp, "/result"), "\\.html$")) == 1)
-})
+  expect_true(length(list.files(paste0(tmp, "/result"), "\\.html$")) == 2)
+  expect_error(create_viewer(audio_dir = paste0(tmp, "/audio"),
+                             picture_dir = paste0(tmp, "/pic"),
+                             table = data.frame(id = 1:2),
+                             output_file = "with_a_map",
+                             output_dir = paste0(tmp, "/result"),
+                             map = TRUE),
+               'If you want to create a map in a viewer, you need to add a glottocode \\(or latitude and longitude\\) column to the datafarame in a table argument\\.')
+  })
 
 unlink(paste0(tmp, "/result"))
 unlink(paste0(tmp, "/audio"))
