### 2020-07-19 6aabaa0
add tests; remove rename_videofiles() till future
