+++ man/rename_videofiles.Rd
@@ -1,42 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rename_videofiles.R
-\name{rename_videofiles}
-\alias{rename_videofiles}
-\title{Rename videofiles}
-\usage{
-rename_videofiles(
-  stimuli,
-  translations = NULL,
-  prefix = NULL,
-  suffix = NULL,
-  order = NULL,
-  path,
-  backup = TRUE
-)
-}
-\arguments{
-\item{stimuli}{character vector of stimuli}
-
-\item{translations}{character vector of translations (optonal)}
-
-\item{prefix}{character vector of length one containing prefix for file names}
-
-\item{suffix}{character vector of length one containing suffix for file names}
-
-\item{order}{numeric vector that define the order of stimuli. By default the
-order of the stimuli is taken.}
-
-\item{path}{path to the directory with videofiles.}
-
-\item{backup}{logical. If TRUE, function creates backup folder with all
-files. By default is TRUE.}
-}
-\value{
-no output
-}
-\description{
-Rename vedeofiles using the template from user.
-}
-\author{
-George Moroz <agricolamz@gmail.com>
-}
