+++ NAMESPACE
@@ -21,7 +21,6 @@ export(get_sound_duration)
 export(get_textgrid_names)
 export(pitch_to_df)
 export(rename_soundfiles)
-export(rename_videofiles)
 export(set_textgrid_names)
 export(srt_to_df)
 export(textgrid_to_df)
