+++ tests/testthat/test-draw-sound.R
@@ -15,25 +15,29 @@ supress_message <- grDevices::dev.off()
 grDevices::png(filename = paste0(tdir, "/draw_sound/test3.png"))
 draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
-           pitch = system.file("extdata", "test.Pitch", package = "phonfieldwork"))
+           pitch = system.file("extdata", "test.Pitch",
+                               package = "phonfieldwork"))
 supress_message <- grDevices::dev.off()
 
 grDevices::png(filename = paste0(tdir, "/draw_sound/test4.png"))
 draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-           pitch = system.file("extdata", "test.Pitch", package = "phonfieldwork"))
+           pitch = system.file("extdata", "test.Pitch",
+                               package = "phonfieldwork"))
 supress_message <- grDevices::dev.off()
 
 grDevices::png(filename = paste0(tdir, "/draw_sound/test5.png"))
 draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
            zoom = c(0.4, 0.65),
-           pitch = system.file("extdata", "test.Pitch", package = "phonfieldwork"))
+           pitch = system.file("extdata", "test.Pitch",
+                               package = "phonfieldwork"))
 supress_message <- grDevices::dev.off()
 
 draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
            zoom = c(0.4, 0.65),
-           pitch = system.file("extdata", "test.Pitch", package = "phonfieldwork"),
+           pitch = system.file("extdata", "test.Pitch",
+                               package = "phonfieldwork"),
            output_file = paste0(tdir, "/draw_sound/test6"))
 
 list.files()
