### 2020-07-19 #385 @agricolamz
Dear @melvidoni,

I have added tests so there is a 82% coverage.

I decided to keep long lines greater then 80 characters when there are links ([here](https://github.com/agricolamz/phonfieldwork/blob/11517a79cdb5b50eb050fb9fe14bf54352083e72/tests/testthat/test-create-viewer.R#L5) is an example).

There is also [one false alarm](https://github.com/agricolamz/phonfieldwork/blob/11517a79cdb5b50eb050fb9fe14bf54352083e72/R/create_viewer.R#L113) about `library()` function: this command is within a character string, so it is not part of the function.
