### 2020-07-20 #385 @agricolamz
Wow, Sign Language specialist! @jonkeane, I'm going to extend my package for working with video for sign languages one day.
