### 2020-07-20 #385 @jonkeane
Hello 👋 I've already started (a very tiny bit) looking around in phonfieldwork, but I am excited to review it! I haven't done sign language research (or really linguistics in general) in a few years now, but when I saw this package in the issues it piqued my curiosity.

If you are thinking about video processing, you might want to checkout [signglossR](https://github.com/borstell/signglossR) which might already have some of that functionality — I'm sure there's space for two packages and the technical needs for sign vs. spoken language phonetics/phonology is different enough it might make sense to keep them separate as well.

Anyway, I'm looking forward to what you've done already 😃
