### 2020-07-21 #385 @agricolamz
Hello! @jonkeane, I knew about this package, but it developed from last time I've seen it!

@melvidoni, if I may to propose my own reviewers? I know it weird since I know them personally, but in case it is not weird:

* Marcelo Araya-Salas, who developed some packages for Bioacoustics like [`warbleR`](https://github.com/maRce10/warbleR), [`Rraven`](https://github.com/maRce10/Rraven) and others
* Niko Partanen, who developed a package for working with the export from [Elan](https://github.com/langdoc/FRelan)
