+++ .github/ISSUE_TEMPLATE/feature-request.md
@@ -1,0 +1,10 @@
+---
+name: Feature request
+about: 'Suggest an idea for this project '
+title: ''
+labels: enhancement
+assignees: ''
+
+---
+
+
