+++ .github/ISSUE_TEMPLATE/feature-request--your-topic.md
@@ -1,10 +1,0 @@
----
-name: 'Feature request: your topic'
-about: Suggest an idea for this project
-title: ''
-labels: enhancement
-assignees: ''
-
----
-
-
