+++ vignettes/phonfieldwork.Rmd
@@ -168,12 +168,33 @@ As a result, a file "first_example.html" was created in the output folder. You c
 
 It is also possible to change the output format, using the `output_format` argument. By dafault it is "html", but you can also use "pptx" (this is a relatively new feature of `rmarkdown`, so update the package in case you get errors). There is also an additional argument `translations`, where you can provide translations for stimuli in order that they appeared near the stimuli on the slide.
 
+It is also possible to use images (or gif, e. g. for a sign language research) as a stimuli. In order to do that you need to provide an absolute or relative path to the file instead of the stimulus and mark in the `external` argument, which of the stimuli is external:
+
+
+```r
+my_image <- system.file("extdata", "r-logo.png", package = "phonfieldwork")
+my_image
+```
+
+```
+## [1] "/home/agricolamz/R/x86_64-pc-linux-gnu-library/4.0/phonfieldwork/extdata/r-logo.png"
+```
+
+```r
+create_presentation(stimuli = c("rzeka", "drzewo", my_image),
+                    external = 3, 
+                    output_dir = getwd())
+```
+
+<iframe src="https://agricolamz.github.io/phonfieldwork/additional/second_example.html" width = 800 height = 600>
+  <p>Your browser does not support iframes :(</p>
+</iframe>
+
 ## Rename collected data
 After collecting data and removing soundfiles with unsuccesful elicitations, one could end up with the following structure:
 
 
 ```
-## ├── s1
 ## │   ├── 01.wav
 ## │   ├── 02.wav
 ## │   └── 03.wav
@@ -181,6 +202,7 @@ After collecting data and removing soundfiles with unsuccesful elicitations, one
 ## │   ├── 01.wav
 ## │   ├── 02.wav
 ## │   └── 03.wav
+## ├── stimuli_presentation.html
 ```
 
 For each speaker `s1` and `s2` there is a folder that containes three audiofiles. Now let's rename the files.
@@ -196,7 +218,6 @@ As a result, you obtain the following structure:
 
 
 ```
-## ├── s1
 ## │   ├── backup
 ## │   │   ├── 01.wav
 ## │   │   ├── 02.wav
@@ -208,6 +229,7 @@ As a result, you obtain the following structure:
 ## │   ├── 01.wav
 ## │   ├── 02.wav
 ## │   └── 03.wav
+## ├── stimuli_presentation.html
 ```
 
 The `rename_soundfiles()` function created a backup folder with all of the unrenamed files, and renamed all files using the prefix provided in the `prefix` argument. There is an additional argument `backup` that can be set to `FALSE` (it is `TRUE` by default), in case you are sure that the renaming function will work properly with your files and stimuli, and you do not need a backup of the unrenamed files.
@@ -223,7 +245,6 @@ rename_soundfiles(stimuli = my_stimuli_df$stimuli,
 
 
 ```
-## ├── s1
 ## │   ├── backup
 ## │   │   ├── 01.wav
 ## │   │   ├── 02.wav
@@ -235,6 +256,7 @@ rename_soundfiles(stimuli = my_stimuli_df$stimuli,
 ## │   ├── s2_tap_2.wav
 ## │   ├── s2_tip_1.wav
 ## │   └── s2_top_3.wav
+## ├── stimuli_presentation.html
 ```
 
[…]
