### 2020-07-21 c0f8ec3
correct create_presentation()
