### 2020-07-21 0524bcd
add some NEWs
