+++ vignettes/phonfieldwork.Rmd
@@ -398,6 +398,7 @@ extract_intervals(file_name = "s1/s1_all.wav",
 
 
 ```
+## ├── s1
 ## │   ├── backup
 ## │   │   ├── 01.wav
 ## │   │   ├── 02.wav
