+++ tests/testthat/test-flextext-to-df.R
@@ -1,7 +1,11 @@
 library(phonfieldwork)
 context("Tests for 'flextext_to_df()' functions")
-file_name <- "https://raw.githubusercontent.com/agricolamz/phonfieldwork/master/vignettes/files/zilo_test.flextext"
-df <- flextext_to_df(file_name)
-test_that("flextext_to_df", {
-  expect_true(nrow(df) == 1170)
+test_that("flextext_to_df_general", {
+  skip_on_cran()
+  file_name <- "https://raw.githubusercontent.com/agricolamz/phonfieldwork/master/vignettes/files/zilo_test.flextext"
+  df <- flextext_to_df(file_name)
+  test_that("flextext_to_df", {
+    skip_on_cran()
+    expect_true(nrow(df) == 1170)
+  })
 })
