+++ tests/testthat/test-draw-sound.R
@@ -1,50 +1,55 @@
 library(phonfieldwork)
 context("Tests for 'draw_sound()' functions")
-tdir <- tempdir()
-dir.create(paste0(tdir, "/draw_sound"))
-grDevices::png(filename = paste0(tdir, "/draw_sound/test1.png"))
-draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"))
-supress_message <- grDevices::dev.off()
 
-grDevices::png(filename = paste0(tdir, "/draw_sound/test2.png"))
-draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-           system.file("extdata", "test.TextGrid",
-                       package = "phonfieldwork"))
-supress_message <- grDevices::dev.off()
+test_that("draw_sound_general", {
+  skip_on_cran()
+  tdir <- tempdir()
+  dir.create(paste0(tdir, "/draw_sound"))
+  grDevices::png(filename = paste0(tdir, "/draw_sound/test1.png"))
+  draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"))
+  supress_message <- grDevices::dev.off()
 
-grDevices::png(filename = paste0(tdir, "/draw_sound/test3.png"))
-draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-           system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
-           pitch = system.file("extdata", "test.Pitch",
-                               package = "phonfieldwork"))
-supress_message <- grDevices::dev.off()
+  grDevices::png(filename = paste0(tdir, "/draw_sound/test2.png"))
+  draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+             system.file("extdata", "test.TextGrid",
+                         package = "phonfieldwork"))
+  supress_message <- grDevices::dev.off()
 
-grDevices::png(filename = paste0(tdir, "/draw_sound/test4.png"))
-draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-           pitch = system.file("extdata", "test.Pitch",
-                               package = "phonfieldwork"))
-supress_message <- grDevices::dev.off()
+  grDevices::png(filename = paste0(tdir, "/draw_sound/test3.png"))
+  draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+             system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
+             pitch = system.file("extdata", "test.Pitch",
+                                 package = "phonfieldwork"))
+  supress_message <- grDevices::dev.off()
 
-grDevices::png(filename = paste0(tdir, "/draw_sound/test5.png"))
-draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-           system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
-           zoom = c(0.4, 0.65),
-           pitch = system.file("extdata", "test.Pitch",
-                               package = "phonfieldwork"))
-supress_message <- grDevices::dev.off()
+  grDevices::png(filename = paste0(tdir, "/draw_sound/test4.png"))
+  draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+             pitch = system.file("extdata", "test.Pitch",
+                                 package = "phonfieldwork"))
+  supress_message <- grDevices::dev.off()
 
-draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-           system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
-           zoom = c(0.4, 0.65),
-           pitch = system.file("extdata", "test.Pitch",
[…]

[…]
