+++ man/intensity_to_df.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/intensity_to_df.R
+\name{intensity_to_df}
+\alias{intensity_to_df}
+\title{Praat Intensity tier to dataframe}
+\usage{
+intensity_to_df(file_name, encoding = "unknown")
+}
+\arguments{
+\item{file_name}{string with a filename or path to the Intensity tier}
+
+\item{encoding}{Intensity tier encoding. Import from \code{readLines()} function.}
+}
+\value{
+a dataframe with columns:  \code{time_start}, \code{time_end},
+\code{Intensity}
+}
+\description{
+Convert a Praat Intensity tier to a dataframe.
+}
+\examples{
+intensity_to_df(system.file("extdata", "test.Intensity", package = "phonfieldwork"))
+
+}
+\author{
+George Moroz <agricolamz@gmail.com>
+}
