+++ NAMESPACE
@@ -19,6 +19,7 @@ export(extract_intervals)
 export(flextext_to_df)
 export(get_sound_duration)
 export(get_textgrid_names)
+export(intensity_to_df)
 export(pitch_to_df)
 export(rename_soundfiles)
 export(set_textgrid_names)
