+++ R/draw_sound.R
@@ -32,11 +32,14 @@
 #' appropriate. At the moment, the only windows that require parameters are the
 #' Kaiser and Gaussian windows. By default, these are set to 2 for kaiser and
 #' 0.4 for gaussian windows.
-#' @param pitch path to the Praat `.Pitch` file or result of pitch_to_df()
-#' function. This variable provide data for visualisation a pitch contour
-#' exported from Praat.
+#' @param pitch path to the Praat `.Pitch` file or result of
+#' \code{pitch_to_df()} function. This variable provide data for visualisation
+#' of a pitch contour exported from Praat.
 #' @param pitch_range vector with the range of frequencies to be displayed.
 #' By default this is set to 75-350 Hz.
+#' @param intensity path to the Praat `.Intensity` file or result of
+#' \code{intensity_to_df()} function. This variable provide data for
+#' visualisation of an intensity contour exported from Praat.
 #' @param output_file the name of the output file
 #' @param output_width the width of the device
 #' @param output_height the height of the device
@@ -72,11 +75,18 @@
 #'            system.file("extdata", "test.TextGrid",
 #'                        package = "phonfieldwork"))
 #'
-# draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
-#            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
-#            pitch = system.file("extdata", "test.Pitch",
-#                                package = "phonfieldwork"),
-#            pitch_range = c(50, 200))
+#' draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+#'            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
+#'            pitch = system.file("extdata", "test.Pitch",
+#'                                package = "phonfieldwork"),
+#'            pitch_range = c(50, 200))
+#' draw_sound(system.file("extdata", "test.wav", package = "phonfieldwork"),
+#'            system.file("extdata", "test.TextGrid", package = "phonfieldwork"),
+#'            pitch = system.file("extdata", "test.Pitch",
+#'                                package = "phonfieldwork"),
+#'            pitch_range = c(50, 200),
+#'            intensity = intensity_to_df(system.file("extdata", "test.Intensity",
+#'                                                    package = "phonfieldwork")))
 #' }
 #' @export
 #'
@@ -116,6 +126,7 @@ draw_sound <- function(file_name,
                        raven_annotation = NULL,
                        pitch = NULL,
                        pitch_range = c(75, 350),
+                       intensity = NULL,
                        output_width = 750,
                        output_height = 500,
                        output_units = "px",
@@ -205,13 +216,13 @@ draw_sound <- function(file_name,
       }
 # plot spectrogram --------------------------------------------------------
 
-      if(is.null(annotation) & is.null(pitch)){
+      if(is.null(annotation) & is.null(pitch) & is.null(intensity)){
         low_boundary <- 0.1
-      } else if(is.null(annotation) & !is.null(pitch)){
+      } else if(is.null(annotation) & (!is.null(pitch)|!is.null(intensity))){
         low_boundary <- 0.1 + 0.17
-      } else if(!is.null(annotation) & is.null(pitch)){
+      } else if(!is.null(annotation) & is.null(pitch)& is.null(intensity)){
         low_boundary <- 0.1 + 0.17
-      } else if(!is.null(annotation) & !is.null(pitch)){
+      } else if(!is.null(annotation) & (!is.null(pitch)|!is.null(intensity))){
         low_boundary <- 0.1 + 0.17 + 0.17
       }
 
[…]
