+++ man/formant_to_df.Rd
@@ -1,0 +1,27 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/formant_to_df.R
+\name{formant_to_df}
+\alias{formant_to_df}
+\title{Praat Formant object to dataframe}
+\usage{
+formant_to_df(file_name, encoding = "unknown")
+}
+\arguments{
+\item{file_name}{string with a filename or path to the Formant file}
+
+\item{encoding}{Pitch tier encoding. Import from \code{readLines()} function.}
+}
+\value{
+a dataframe with columns:  \code{time_start}, \code{time_end},
+\code{frequency}, \code{bandwidth} and \code{formant}
+}
+\description{
+Convert a Praat Formant object to a dataframe.
+}
+\examples{
+formant_to_df(system.file("extdata", "e.Formant", package = "phonfieldwork"))
+
+}
+\author{
+George Moroz <agricolamz@gmail.com>
+}
