+++ NEWS.md
@@ -1,9 +1,10 @@
 # phonfieldwork 0.0.8
 
 - add `intensity`, `picth` and `pitch_range` arguments to the `draw_sound()` function
+- add the `formant_to_df()` function
 - add the `picth_to_df()` function
 - add the `intensity_to_df()` function
-- add an argument `external` to the `create_presentation` function in order to mark external images or gifs
+- add an argument `external` to the `create_presentation()` function in order to mark external images or gifs
 
 # phonfieldwork 0.0.7
 
