+++ NAMESPACE
@@ -17,6 +17,7 @@ export(eaf_to_df)
 export(exb_to_df)
 export(extract_intervals)
 export(flextext_to_df)
+export(formant_to_df)
 export(get_sound_duration)
 export(get_textgrid_names)
 export(intensity_to_df)
