+++ presentations/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje_moroz_phonfieldwork.Rmd
@@ -23,7 +23,7 @@ editor_options:
 ---
 
 ```{r setup, include=FALSE}
-knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
+knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
 # library(qrcode)
 # png(filename="images/01_qrcode.png", width = 200, height = 200)
 # qrcode_gen("https://github.com/agricolamz/phonfieldwork/raw/master/presentations/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje_moroz_phonfieldwork.pdf")
@@ -47,6 +47,32 @@ knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)
 The `phonfieldwork` package is created for helping with items 3, 4 (partially), 5, 6, and 9.
 
 # Installation of the package
+
+## Install `phonfieldwork`
+
+`phonfieldwork` is an R package, so you need to install [R](https://cloud.r-project.org/), [RStudio](https://rstudio.com/products/rstudio/download/#download) (optional) or use [rstudio.cloud](rstudio.cloud). There are two possibilities for installing package in R:
+
+* official version from CRAN
+
+```{r, eval = FALSE}
+install.packages("phonfieldwork")
+```
+
+* development version from GitHub
+
+```{r, eval = FALSE}
+devtools::install_github("agricolamz/phonfieldwork")
+```
+
+\pause
+
+Since this package is under [rOpenSci review](https://github.com/ropensci/software-review/issues/385) there is a chance that in couple months the adress could be changed to `ropensci/phonfieldwork`, and documentation page will be moved from [agricolamz.github.io/phonfieldwork](agricolamz.github.io/phonfieldwork) to
+[docs.ropensci.org/phonfieldwork](docs.ropensci.org/phonfieldwork).
+
+```{r}
+library("phonfieldwork")
+packageVersion("phonfieldwork") # Unreleased version
+```
 
 # Creating your presentation
 
