+++ presentations/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje_moroz_phonfieldwork.Rmd
@@ -107,9 +107,13 @@ packageVersion("phonfieldwork") # Unreleased version
 
 # Creating a data viewer -- template for the data sharing
 
-# Cite
+# Get help and cite
 
-## Cite the package
+## Get help and cite
+
+You can always write an email or open an [issue on GitHub](https://github.com/agricolamz/phonfieldwork/issues), asking some questions.
+
+The most recent citation information is avalible with this command:
 
 ```{r}
 citation("phonfieldwork")
