+++ presentations/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje_moroz_phonfieldwork.Rmd
@@ -3,14 +3,13 @@ title: "Phonetic fieldwork and experiments with the `phonfieldwork` package for 
 author: "George Moroz"
 institute: "Linguistic Convergence Laboratory, NRU HSE"
 date: |
-    | 08.08.2020, Grupo de estadística para el estudio del lenguaje
+    | 08 August 2020, Grupo de estadística para el estudio del lenguaje
     |
     |
     | Presentation is available here: \alert{tinyurl.com/y6lf5ch4}
     | ![](images/01_qrcode.png)'
 output: 
   beamer_presentation:
-    df_print: kable
     latex_engine: xelatex
     citation_package: natbib
     keep_tex: false
@@ -28,6 +27,13 @@ knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
 # png(filename="images/01_qrcode.png", width = 200, height = 200)
 # qrcode_gen("https://github.com/agricolamz/phonfieldwork/raw/master/presentations/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje_moroz_phonfieldwork.pdf")
 # dev.off()
+
+unlink("data/s1", recursive = TRUE)
+unlink("data/s2", recursive = TRUE)
+dir.create("data/s1")
+dir.create("data/s2")
+file.copy(paste0("../backup/s1/", list.files("../backup/s1/")), "data/s1/")
+file.copy(paste0("../backup/s2/", list.files("../backup/s2/")), "data/s2/")
 ```
 
 # Introduction
@@ -63,6 +69,14 @@ All of these tasks can be solved by a mixture of different tools:
 
 `phonfieldwork` provides a functionality that will make it easier to solve those tasks independently of any additional tools. You can also compare the functionality with other packages: `rPraat` [@borzil16], `textgRid` [@reidy16], `pympi` [@lubbers13]
 
+## Philosophy of the `phonfieldwork` package
+
+* each stimulus is a separate file
+* researcher carefully listens to consultants to make sure that they are producing the kind of speech they wanted
+* in case a speaker does not produce clear repetitions, researcher ask them to repeat the task
+
+There are some phoneticians who prefer to record everything for language documentation purposes. I think that should be a separate task. If you insist on recording everything, it is possible to run two recorders at the same time: one could run during the whole session, while the other is used to produce small audio files. You can also use special software to record your stimuli automatically on a computer (e.g. `PsychoPy` [@peirce19]).
+   
 # Installation of the package
 
 ## Install `phonfieldwork`
@@ -93,7 +107,89 @@ packageVersion("phonfieldwork") # Unreleased version
 
 # Creating your presentation
 
+##   Make a list of your stimuli
+
+There are several ways to enter information about a list of stimuli into R:
+
+* list them with the `c()` command
+
+```{r}
+my_stimuli <- c("tip", "tap", "top")
+```
+
+* read from `.csv` file with the `read.csv()` function:
+
+```{r, eval = FALSE}
+my_stimuli_df <- read.csv("my_stimuli_df.csv")
+```
+
+* read from `.xls` or `.xlsx` file with the `read_xls()` or `read_xlsx()` function from the `readxl` package (run `install.packages("readxl")` in case you don't have it installed):
+
+```{r, eval = FALSE}
+library("readxl")
+my_stimuli_df <- read_xlsx("my_stimuli_df.xlsx")
+```
+
+
+##  Create a presentation based on a list of stimuli
+
 # Data renaming
+
+## Obtained data
+
[…]
