+++ presentations/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estadística_para_el_estudio_del_lenguaje_moroz_phonfieldwork.Rmd
@@ -28,12 +28,13 @@ knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
 # qrcode_gen("https://github.com/agricolamz/phonfieldwork/raw/master/presentations/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje_moroz_phonfieldwork.pdf")
 # dev.off()
 
-unlink("data/s1", recursive = TRUE)
-unlink("data/s2", recursive = TRUE)
+unlink("data", recursive = TRUE)
+dir.create("data")
 dir.create("data/s1")
 dir.create("data/s2")
 file.copy(paste0("../backup/s1/", list.files("../backup/s1/")), "data/s1/")
 file.copy(paste0("../backup/s2/", list.files("../backup/s2/")), "data/s2/")
+file.copy(paste0("../backup/", list.files("../backup/", pattern = "test")), "data/")
 ```
 
 # Introduction
@@ -133,6 +134,28 @@ my_stimuli_df <- read_xlsx("my_stimuli_df.xlsx")
 
 ##  Create a presentation based on a list of stimuli
 
+```{r}
+create_presentation(stimuli = my_stimuli, # it is "tip" "tap" "top"
+                    output_file = "first_example",
+                    output_dir = "data/")
+```
+
+Here is [the result](https://agricolamz.github.io/phonfieldwork/additional/first_example.html). \pause
+
+It is also possible to use images (or `.gif`) as a stimuli:
+
+```{r}
+my_image <- system.file("extdata", "r-logo.png", package = "phonfieldwork")
+my_image
+
+create_presentation(stimuli = c("rzeka", "drzewo", my_image),
+                    external = 3,
+                    output_file = "second_example",
+                    output_dir = "data/")
+```
+
+Here is [the result](https://agricolamz.github.io/phonfieldwork/additional/second_example.html).
+
 # Data renaming
 
 ## Obtained data
@@ -140,7 +163,7 @@ my_stimuli_df <- read_xlsx("my_stimuli_df.xlsx")
 After collecting data and removing soundfiles with unsuccesful elicitations, one could end up with the following structure:
 
 ```{bash, echo = FALSE}
-tree data | tail -n 10 | head -n 8
+tree data | tail -n 11 | head -n 8
 ```
 
 ## Rename collected data
@@ -152,7 +175,7 @@ rename_soundfiles(stimuli = my_stimuli, # it is "tip" "tap" "top"
 ```
 
 ```{bash, echo = FALSE}
-tree data | tail -n 14 | head -n 12
+tree data | tail -n 15 | head -n 12
 ```
 
 ## Rename collected data
@@ -166,7 +189,7 @@ rename_soundfiles(stimuli = my_stimuli, # it is "tip" "tap" "top"
 ```
 
 ```{bash, echo = FALSE}
-tree data | tail -n 14 | head -n 12
+tree data | tail -n 15 | head -n 12
 ```
 
 ## Rename collected data
@@ -193,15 +216,219 @@ get_sound_duration(sounds_from_folder = "data/s2/")
 
 # Data merging
 
+## Merge all data together
+
+After all the files are renamed, you can merge them into one:
+
+```{r}
+concatenate_soundfiles(path = "data/s1/",
+                       result_file_name = "s1_all")
+```
+
+```{bash, echo = FALSE}
+tree data/s1 | head -n 10
+```
+
+## Merge all data together
+
+```{r, echo=FALSE}
+draw_sound("data/s1/s1_all.wav", "data/s1/s1_all.TextGrid", title = "s1_all")
+```
+
 # Data annotation
 
+## Annotate your data
+
[…]
