### 2020-08-20 agricolamz/phonfieldwork#26 @agricolamz
`create_glossed_document()` not only for flextext
It is nice to think about creation of a document from any type of annotation.
