+++ vignettes/phonfieldwork.Rmd
@@ -2,7 +2,7 @@
 title: "Phonetic fieldwork and experiments with `phonfieldwork` package"
 author: "G. Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
 bibliography: bibliography.bib
-date: "2020-08-02"
+date: "2020-08-20"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Phonetic fieldwork and experiments with `phonfieldwork` package}
@@ -31,8 +31,8 @@ This command will install the last stable version of the `phonfieldwork` package
 
 
 ```r
-install.packages("devtools")
-devtools::install_github("agricolamz/phonfieldwork")
+install.packages("remotes")
+remotes::install_github("agricolamz/phonfieldwork")
 ```
 
 If you have any trouble installing the package, you will not be able to use its functionality. In that case you can create [an issue on Github](https://github.com/agricolamz/phonfieldwork/issues) or send an email. Since this package could completely destroy your data, **please do not use it until you are sure that you have made a backup**.
@@ -62,8 +62,8 @@ citation("phonfieldwork")
 
 ```
 ## 
-## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork package_. <URL:
-## https://CRAN.R-project.org/package=phonfieldwork>.
+## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork
+## package_. <URL: https://CRAN.R-project.org/package=phonfieldwork>.
 ## 
 ## A BibTeX entry for LaTeX users is
 ## 
@@ -634,15 +634,24 @@ exb_to_df(system.file("extdata", "test.exb", package = "phonfieldwork"))
 ```
 
 ```
-##   tier id content tier_name tier_type tier_category tier_speaker time_start  time_end   source
-## 3    1  1       t     X [v]         t             v         SPK0 0.06908955 0.2498984 test.exb
-## 1    1  2       e     X [v]         t             v         SPK0 0.24989836 0.3807275 test.exb
-## 5    1  3       s     X [v]         t             v         SPK0 0.38072750 0.4042473 test.exb
-## 7    1  4       t     X [v]         t             v         SPK0 0.40424735 0.6526757 test.exb
-## 4    2  1       C     X [v]         a             v         SPK0 0.06908955 0.2498984 test.exb
-## 2    2  2       V     X [v]         a             v         SPK0 0.24989836 0.3807275 test.exb
-## 6    2  3       C     X [v]         a             v         SPK0 0.38072750 0.4042473 test.exb
-## 8    2  4       C     X [v]         a             v         SPK0 0.40424735 0.6526757 test.exb
+##   tier id content tier_name tier_type tier_category tier_speaker time_start
+## 3    1  1       t     X [v]         t             v         SPK0 0.06908955
+## 1    1  2       e     X [v]         t             v         SPK0 0.24989836
+## 5    1  3       s     X [v]         t             v         SPK0 0.38072750
+## 7    1  4       t     X [v]         t             v         SPK0 0.40424735
[…]
