+++ R/textgrid_to_df.R
@@ -12,7 +12,7 @@
 #'
 #' @return a dataframe with columns:  \code{id}, \code{time_start},
 #' \code{time_end} (if it is an interval tier -- the same as the start value),
-#' \code{content}, \code{tier} and \code{source}
+#' \code{content}, \code{tier}, \code{tier_name} and \code{source}
 #'
 #' @examples
 #' textgrid_to_df(system.file("extdata", "test.TextGrid",
@@ -29,20 +29,45 @@ textgrid_to_df <- function(file_name,
     } else{
       tg <- readLines(file_name, encoding = encoding)
     }
-    n_tiers <- as.double(regmatches(tg[7], regexpr("\\d{1,}", tg[7])))
-    lapply(1:n_tiers, function(x){
-      df <- phonfieldwork::tier_to_df(tg, x)
-      df$tier <- x
-      return(df)
-    }) ->
+
+    sum(grepl("item ?\\[\\d{1,}\\]:", tg)) < 1
+
+    lapply(split(seq_along(tg),
+                 cumsum(grepl("item ?\\[\\d{1,}\\]:", tg)))[-1],
+           function(i){
+             class <- unlist(strsplit(tg[i[2]], '"'))[2]
+             step_by = ifelse(class == "IntervalTier", 4, 3)
+             start_max = ifelse(class == "IntervalTier", 9, 8)
+             data.frame(
+               id = seq_along(seq(8, length(i), by = step_by)),
+               time_start = gsub("[^0-9.]", "", tg[i[seq(8, length(i),
+                                                         by = step_by)]]),
+               time_end = gsub("[^0-9.]", "", tg[i[seq(start_max, length(i),
+                                                       by = step_by)]]),
+               content = tg[i[seq(start_max+1, length(i), by = step_by)]],
+               tier = gsub("[^0-9]", "", tg[i[1]]),
+               tier_name = unlist(strsplit(tg[i[3]], '"'))[2],
+               stringsAsFactors = FALSE)
+           }) ->
       l
+
     result <- do.call(rbind, l)
+
+
+    result$content <- unlist(lapply(result$content, function(j){
+      unlist(strsplit(j, '"'))[2]
+    }))
+    result$time_start <- as.numeric(result$time_start)
+    result$time_end <- as.numeric(result$time_end)
+    result$tier <- as.numeric(result$tier)
     if(grepl("TextGrid", file_name[2])){
       source <- "custom_file"
     } else{
       source <- unlist(strsplit(normalizePath(file_name), "/"))
     }
     result$source <- source[length(source)]
+
+    rownames(result) <- NULL
     return(result[order(result$time_start),])
   } else {
     files <- paste0(normalizePath(textgrids_from_folder),
