+++ vignettes/phonfieldwork.Rmd
@@ -2,7 +2,7 @@
 title: "Phonetic fieldwork and experiments with `phonfieldwork` package"
 author: "G. Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
 bibliography: bibliography.bib
-date: "2020-08-20"
+date: "2020-08-22"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Phonetic fieldwork and experiments with `phonfieldwork` package}
@@ -62,8 +62,8 @@ citation("phonfieldwork")
 
 ```
 ## 
-## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork
-## package_. <URL: https://CRAN.R-project.org/package=phonfieldwork>.
+## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork package_. <URL:
+## https://CRAN.R-project.org/package=phonfieldwork>.
 ## 
 ## A BibTeX entry for LaTeX users is
 ## 
@@ -584,21 +584,21 @@ textgrid_to_df(system.file("extdata", "test.TextGrid", package = "phonfieldwork"
 ```
 
 ```
-##    id time_start   time_end content tier        source
-## 1   1 0.00000000 0.01246583            1 test.TextGrid
-## 6   1 0.00000000 0.01246583            2 test.TextGrid
-## 2   2 0.01246583 0.24781914       t    1 test.TextGrid
-## 7   2 0.01246583 0.24781914            2 test.TextGrid
-## 11  1 0.01246583 0.01246583       t    3 test.TextGrid
-## 3   3 0.24781914 0.39552363       e    1 test.TextGrid
-## 8   3 0.24781914 0.39552363            2 test.TextGrid
-## 12  2 0.24781914 0.24781914       e    3 test.TextGrid
-## 4   4 0.39552363 0.51157715       s    1 test.TextGrid
-## 9   4 0.39552363 0.51157715            2 test.TextGrid
-## 13  3 0.39552363 0.39552363       s    3 test.TextGrid
-## 5   5 0.51157715 0.65267574       t    1 test.TextGrid
-## 10  5 0.51157715 0.65267574            2 test.TextGrid
-## 14  4 0.51157715 0.51157715       t    3 test.TextGrid
+##    id time_start   time_end content tier       tier_name        source
+## 1   1 0.00000000 0.01246583            1       intervals test.TextGrid
+## 6   1 0.00000000 0.01246583            2 empty_intervals test.TextGrid
+## 2   2 0.01246583 0.24781914       t    1       intervals test.TextGrid
+## 7   2 0.01246583 0.24781914            2 empty_intervals test.TextGrid
+## 11  1 0.01246583 0.01246583       t    3          points test.TextGrid
+## 3   3 0.24781914 0.39552363       e    1       intervals test.TextGrid
+## 8   3 0.24781914 0.39552363            2 empty_intervals test.TextGrid
+## 12  2 0.24781914 0.24781914       e    3          points test.TextGrid
+## 4   4 0.39552363 0.51157715       s    1       intervals test.TextGrid
+## 9   4 0.39552363 0.51157715            2 empty_intervals test.TextGrid
+## 13  3 0.39552363 0.39552363       s    3          points test.TextGrid
+## 5   5 0.51157715 0.65267574       t    1       intervals test.TextGrid
+## 10  5 0.51157715 0.65267574            2 empty_intervals test.TextGrid
[…]
