+++ man/textgrid_to_df.Rd
@@ -27,6 +27,9 @@ Convert Praat TextGrid to a dataframe.
 textgrid_to_df(system.file("extdata", "test.TextGrid",
                            package = "phonfieldwork"))
 
+# this is and example of reading a short .TextGrid format
+textgrid_to_df(system.file("extdata", "test_short.TextGrid",
+                           package = "phonfieldwork"))
 }
 \author{
 George Moroz <agricolamz@gmail.com>
