+++ R/annotate_textgrid.R
@@ -12,7 +12,6 @@
 #' @param each non-negative integer. Each element of x is repeated each times
 #' @param backup logical. If TRUE (by default) it creates a backup tier.
 #' @param write logical. If TRUE (by dafault) it overwrites an existing tier.
-#' @param encoding TextGrid encoding. Import from \code{readLines()} function.
 #'
 #' @return a string that contain TextGrid. If argument write is \code{TRUE},
 #' then no output.
@@ -26,20 +25,24 @@
 #'
 #' @export
 #'
+#' @importFrom uchardet detect_file_enc
+#'
 
 annotate_textgrid <- function(annotation,
                               textgrid,
                               tier = 1,
                               each = 1,
                               backup = TRUE,
-                              write = TRUE,
-                              encoding = "unknown") {
+                              write = TRUE) {
 
 # read TextGrid -----------------------------------------------------------
   if(grepl("TextGrid", textgrid[2])){
     tg <- textgrid
   } else{
-    tg <- readLines(normalizePath(textgrid), encoding = encoding)
+    # thanks to Artem Klevtsov for this code
+    con <- file(textgrid, encoding = uchardet::detect_file_enc(textgrid))
+    tg <- readLines(con)
+    close(con)
   }
 
 # get start and end info about tiers --------------------------------------
