+++ R/tier_to_df.R
@@ -7,7 +7,6 @@
 #' @param file_name string with a filename or path to the TextGrid
 #' @param tier value that could be either ordinal number of the tier either
 #' name of the tier. By default is '1'.
-#' @param encoding TextGrid encoding. Import from \code{readLines()} function.
 #'
 #' @return a dataframe with columns:  \code{id}, \code{time_start},
 #' \code{time_end}, \code{content}, , \code{tier_name}
@@ -21,9 +20,8 @@
 #' @export
 #'
 
-tier_to_df <- function(file_name, tier = 1, encoding = "unknown"){
-  df <- phonfieldwork::textgrid_to_df(file_name = file_name,
-                                      encoding = encoding)
+tier_to_df <- function(file_name, tier = 1){
+  df <- phonfieldwork::textgrid_to_df(file_name = file_name)
 
   if(is.numeric(tier)){
     if(tier > max(df$tier)){
