+++ R/get_textgrid_names.R
@@ -5,7 +5,6 @@
 #' @author George Moroz <agricolamz@gmail.com>
 #'
 #' @param textgrid path to the TextGrid
-#' @param encoding TextGrid encoding. Import from \code{readLines()} function.
 #'
 #' @return return a vector of tier names from given TextGrid
 #' @examples
@@ -15,18 +14,6 @@
 #' @export
 #'
 
-get_textgrid_names <- function(textgrid, encoding = "unknown"){
-# read TextGrid -----------------------------------------------------------
-  if(grepl("TextGrid", textgrid[2])){
-    tg <- textgrid
-  } else{
-    tg <- readLines(normalizePath(textgrid), encoding = encoding)
-  }
-
-# extract names of tiers --------------------------------------------------
-  lapply(grep('name = ".*"', tg), function(i){
-    unlist(strsplit(tg[i], '"'))[2]
-  }) ->
-    names
-  return(unlist(names))
+get_textgrid_names <- function(textgrid){
+  unique(unlist(textgrid_to_df(textgrid)$tier_name))
 }

[…]
