### 2020-08-28 #385 @nikopartanen
- I have met George Moroz many times. I'm familiar with his work, and have used his another R package lingtypology often in my own work.
- [x] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work

---

`phonfieldwork` is a very interesting and useful R package, and is a very welcome addition to the tools we have available to work with speech data. I also did this review during a few weeks in several times, please let me know if some of my comments relate to something already outdated. I installed the package again from GitHub before submitting my review and tried to check that my comments are still valid.

How I understand it, the package has several larger areas where it can be used, and which form the pipeline described also in the documentation. 1) Creating stimuli sets for data collection, 2) Organising data collected with these stimuli, 3) Processing and further annotating the annotation files, 4) Reading into R resulting datasets, and 5) Visualizing this data.

The package is designed so that the different part of the pipeline work well together. Many of my recommendations relate to the ways how the package could be made more useful also in a little bit different scenarios, where the data is collected differently or the user already has some materials in a different format and wants to take advantage of functionalities of the package later in the pipeline.

#### Creating stimuli and organizing the files

This functionality is very useful and well thought. There aren't that many tools around for this kind of task. In the package documentation there is a good explanation of the intended workflow, with notes about recording practices.

I recorded some test audio with Audacity following the documentation, and ended up with files called `1.wav`, `2.wav` and `3.wav`, containing words *tip*, *tap* and *top*. I used these in most of my testing within this part and visualizations.

One problem I encountered with `rename_soundfiles` function is that there doesn't seem to be any way to keep the original recording order intact. I end up with files `tap.wav`, `tip.wav` and `top.wav`. However, even when I have the backup, I don't have the information stored anywhere that `1.wav` is now called `tip.wav`. It could be useful to have an argument that would rename the files, but add a number following the original order, i.e. `01_tip.wav`, `02_tap.wav` and `03_top.wav`. Or maybe the function could also write a log file into the backup directory that lists how the renaming was done. I'm just thinking about a scenario where we realize much later that something has gone wrong, and for whatever reason we need to go back to original recordings. As far as I understand, now also `suffix = paste0("_", 1:3)` would not keep the original stimuli order? I think one could often want to record the same word under a somehow different stimulus, and maybe for analysis it would be clearer to keep the original order.

I understand from the documentation that the user creates the stimuli, shows them to a research participant, records stimuli as individual clips, and then ends up with a list of recordings that is the length of stimuli list, and has thereby already removed accidental double takes. This if fine, and in many cases a really well functioning workflow, and the package does provide a good help in doing this.

It can be useful to take a look into [SpeechRecorder](https://www.bas.uni-muenchen.de/Bas/software/speechrecorder/) by [BAS](https://www.bas.uni-muenchen.de/), which provides similar functionality. The advantage of this software is that one can record directly within a template, and use a microphone through an audio interface connected into a laptop. The main problem I see in using `phonfieldwork`'s stimuli list is that it isn't entirely clear from the documentation how the researcher should handle repeated or missed elicitations, and when the number of elicitations is in hundreds of items, then managing this can be really complicated. This is of course a problem that documentation can solve, and there are already good recommendations related to the recordings.

Maybe it would be useful to add into a stimuli list a clear sequential id number, so the user can easily listen, for example, to the stimulus 75 and check that this is the same as 75th file in their recording directory. I understand the page number in stimuli file probably already fills this need, and maybe it is enough. Additionally, one could also specify in the `rename_soundfiles` new arguments `missing`, where the user could basically be able to say that "for whatever reason we don't have recordings for stimulus 21, 34 and 94, rename things accordingly". Otherwise the user gets the error `Number of files is smaller then number of stimuli`.

#### Processing data

Annotating TextGrid files works very well, and the functions are well thought and form a clear pipeline. Examples in the documentation are also clear.

It seems to me that there is no way to create a TextGrid file besides `concatenate_soundfiles` function. I think, however, that in many situations we have annotations in some of the formats the package can read, and one would like to work further with the package's spectrogram visualization methods. When we want to edit the annotations at this level of exactness, having it in Praat is very beneficial. So possibly it could be useful if there was a function `create_textgrid` that would create an empty TextGrid file that the user could populate with the functions `annotate_textgrid` and `create_subannotation`. Here I'm thinking a situation where the user would like to jump into the `phonfieldwork`'s pipeline more in the middle, i.e. writing intermediate Praat files from segments read from another corpus.

Separating the folder processing side of `concatenate_soundfiles` and working with individual files could be good. Maybe user could also provide a list of audio files to be concatenated? It is possible this wouldn't fit to the logic of the pipeline the package is designed for, but I leave the idea here just in case.

#### Reading data into R

Reading datasets into R works really well. I tested extensively with both ELAN and Exmaralda corpora, and the resulting data frame is well structured, and the package works without problems with different ELAN files. I must emphasize that these functions worked better with .eaf and .exb files than anything I have seen before, so I'm very happy with them.

I used while testing Exmaralda-files from [Nganasan Spoken Language Corpus](https://corpora.uni-hamburg.de/hzsk/de/islandora/object/spoken-corpus:nslc-0.2) and [Selkup Language Corpus](https://corpora.uni-hamburg.de/hzsk/de/islandora/object/spoken-corpus:slc-1.0.0). I also used an individual FlexText file I found from [here](https://osf.io/dxtne/), which belongs to corpus [The Documentation of Eibela: An archive of Eibela language materials from the Bosavi region](https://elar.soas.ac.uk/Collection/MPI1013856). For testing ELAN I used various files I had at my laptop, as this is very common format it was easier to find different examples. Reading Using ELAN files was also entirely free from adventures.

With some Exmaralda files I did encounter problems, this should repeat my error:

download.file(url = "https://corpora.uni-hamburg.de/hzsk/de/islandora/object/transcript:slc-1.0.0_BEP_1973_Itja1_flk/datastream/EXB/BEP_1973_Itja1_flk.exb", destfile = "BEP_1973_Itja1_flk.exb")

phonfieldwork::exb_to_df("BEP_1973_Itja1_flk.exb")
```

Which results in:

```
Error in data.frame(tier = i, id = seq_along(content), content = content,  :
  arguments imply differing number of rows: 1, 0
```

Maybe there could be some generic error message that prints the path of the file and tells the file cannot be parsed, and then shows the underlying error message. In a typical corpus creation workflow the files we fail to parse are usually a very good indicator for the files that need to be checked. So that some files fail to parse is just typical, and no problem necessarily in itself. Especially if the file is broken it is better to fail than reading it wrong. I assume in some corpora there are just very unusual file structures, possibly related to old conversions from a format to format, so it is impossible to predict everything that is inside those files. But there may be some patterns that are so common that checking what is going on can improve the function.

I also encountered one FlexText file that I couldn't parse, which is [here](https://osf.io/xcafp/). This file doesn't seem to contain the word-level annotations, so that it doesn't work is not surprising, of course. If this kind of files are more common (I have no ideas, FlexText files are surprisingly hard to find!), it could help to have a specific error or warning message for this.

I think most of the time after the user reads these files into R, they would need to filter it a bit and probably use `pivot_wider` to get into right columns the content they want to work with. This isn't strictly speaking a matter of this package, but it could be nice to have in a vignette an example of such process, especially if the goal is to make the package beginner friendly. Something like:

download.file("https://corpora.uni-hamburg.de/hzsk/de/islandora/object/transcript:nslc-0.2_ChKD_72_ManyTents_flks/datastream/EXB/ChKD_72_ManyTents_flks.exb", destfile = "ChKD_72_ManyTents_flks.exb")

phonfieldwork::exb_to_df(file_name = "ChKD_72_ManyTents_flks.exb") %>%
    filter(tier_category %in% c("tx", "ps", "mp", "mc", "mb", "ge", "gr")) %>%
    select(-tier, -id, -tier_name, -tier_type) %>%
    pivot_wider(names_from = tier_category, values_from = c(content))
```

This way we end up with a nice data frame that has tokens, POS-tags and morphological analysis organized so that we have one observation (token now, could also be the utterance) per one row. It is impossible, of course, to predict which tiers the user wants to work with, which is why simple examples that give a direction to further possibilities could be useful. Or just to mention that `pivot_wider` is something to look into for further processing and analysis of the resulting data frame.

**Note: These corpora are distributed under CC BY-NC-SA license, so using some other examples in the package vignette is certainly recommendable, as the license is not compatible with that of the package. I assume using these files in my review and testing is without problems. I originally wanted to include a screenshot of the resulting dataframe, but I excluded it now in order not to have actual corpus content shown.**

I agree with the other reviewer that reading individual files and reading all the files in the directory should be somehow separated. Especially since when the user reads in whole directories there should be good control over things like finding the files recursively or not etc.

#### Visualizations

These work well, and as described. I'm also personally extremely happy to be able to create good-looking spectrograms in R. These are just small suggestions that possibly could improve the usability.

Adding the annotations from a data frame into a visualization made with `draw_sound` works really beautifully. It took me quite much time to understand that it is possible to pass a dataframe into `annotations` argument. Now the help text says `a source for annotation files (e. g. TextGrid)` – a mention of a data frame should also be here, this is now too easy to miss, so it should be prominently in the documentation as well.

As a technical note, using a tibble instead a data.frame seems to give an error here. I think it would be important that the package would not make a difference between data.frame and tbl_df classes.

Bonus points for Raven style annotations, that's a great addition.

#### Further comments

I think `create_glossed_document` is a very useful function, but I'm not entirely sure if this package is ultimately the best place for it. It's a bit aside from the current pipeline the package provides, although it is a really important tool. One way to connect it better would be to be able to give it instead of a FlexText file a data frame with specific columns. Then it would link more clearly to other capablities of the package: reading the corpus, manipulating and analyzing it in R, and then converting some specific parts to glossed documents that can be used elsewhere.

What it comes to the HTML-export, it could be worth checking out [Leipzig.js](https://bdchauvette.net/leipzig.js/) JavaScript library, as the glosses are much more pretty when aligned correctly. However, especially Word export is something I've heard wished for a lot, so it's nice to see it exists and works well here.
