+++ ignore/review_answer_draft.md
@@ -1,0 +1,8 @@
+# For  @jonkeane
+
+## Think about making classes
+
+I keep it in mind, but for the time being I created a function `read_textgrid()` that is more general then previous, since it deals with encodings.
+
+
+# For  @nikopartanen
