### 2020-09-27 16fc950
add autonumber and logging arguments to  rename_soundfile()
