+++ tests/testthat/test-rename-soundfiles.R
@@ -9,7 +9,7 @@ rename_soundfiles(stimuli = c("s1", "s2"),
                   path = paste0(tdir, "/rename_audio"))
 test_that("rename_soundfiles", {
   expect_equal(list.files(paste0(tdir, "/rename_audio"), "\\.wav$"),
-               c("s1.wav", "s2.wav"))
+               c("1_s1.wav", "2_s2.wav"))
 })
 unlink(paste0(tdir, "/rename_audio"))
 rm(tdir)
