+++ ignore/review_answer_draft.md
@@ -2,7 +2,10 @@
 
 ## Think about making classes
 
-I keep it in mind, but for the time being I created a function `read_textgrid()` that is more general then previous, since it deals with encodings.
+I will keep it in mind, but for the time being I created a function `read_textgrid()` that is more general then previous, since it deals with encodings.
 
 
 # For  @nikopartanen
+
+## Creating stimuli and organizing the files
+I added `autonumber` and `loging` arguments to `rename_soundfiles()` function. They both can be switched off. The `autonumber` argument add prefix with number (and appropriate zero padding). The `logging` arguments creates a `logging.csv` file with the change correspondencies.
