+++ NEWS.md
@@ -7,6 +7,7 @@
 - add the `intensity_to_df()` function
 - add an argument `external` to the `create_presentation()` function in order to mark external images or gifs
 - remove all `encoding` arguments and replace it with encoding autodetection from `uchardet` (thanks to Artem Klevtsov for help)
+- add `autonumber` and `loging` arguments to `rename_soundfiles()` function (thanks to Niko Partanen)
 
 # phonfieldwork 0.0.7
 
