### 2020-09-27 69e9094
small code style changes and styler
