+++ R/eaf_to_df.R
@@ -14,7 +14,6 @@
 #'
 #' @examples
 #' eaf_to_df(system.file("extdata", "test.eaf", package = "phonfieldwork"))
-#'
 #' @export
 #' @importFrom xml2 read_xml
 #' @importFrom xml2 xml_find_all
@@ -22,8 +21,8 @@
 #' @importFrom xml2 xml_text
 #' @importFrom xml2 xml_children
 
-eaf_to_df <- function(file_name, eafs_from_folder = NULL){
-  if(is.null(eafs_from_folder)){
+eaf_to_df <- function(file_name, eafs_from_folder = NULL) {
+  if (is.null(eafs_from_folder)) {
     # read file
     l <- xml2::read_xml(file_name)
     # extract tiers
@@ -34,31 +33,43 @@ eaf_to_df <- function(file_name, eafs_from_folder = NULL){
     tier_types <- xml2::xml_attr(t, "LINGUISTIC_TYPE_REF")
 
     # create list of dataframes
-    lapply(seq_along(t), function(i){
-      content <- xml2::xml_text(xml2::xml_find_all(t[[i]],
-                                              "ANNOTATION/*/ANNOTATION_VALUE"))
-      ts1 <- xml2::xml_attr(xml2::xml_children(xml2::xml_children(t[[i]])),
-                            "TIME_SLOT_REF1")
-      ts2 <- xml2::xml_attr(xml2::xml_children(xml2::xml_children(t[[i]])),
-                            "TIME_SLOT_REF2")
-      a_id <- xml2::xml_attr(xml2::xml_children(xml2::xml_children(t[[i]])),
-                             "ANNOTATION_ID")
-      ar <- xml2::xml_attr(xml2::xml_children(xml2::xml_children(t[[i]])),
-                           "ANNOTATION_REF")
-      if(length(content) > 0){
-        data.frame(tier = i,
-                   id = seq_along(content),
-                   content = content,
-                   tier_name = tier_names[i],
-                   tier_type = tier_types[i],
-                   ts_start = ts1,
-                   ts_end = ts2,
-                   a_id = a_id,
-                   ar = ar,
-                   stringsAsFactors = FALSE)
+    r <- lapply(seq_along(t), function(i) {
+      content <- xml2::xml_text(xml2::xml_find_all(
+        t[[i]],
+        "ANNOTATION/*/ANNOTATION_VALUE"
+      ))
+      ts1 <- xml2::xml_attr(
+        xml2::xml_children(xml2::xml_children(t[[i]])),
+        "TIME_SLOT_REF1"
+      )
+      ts2 <- xml2::xml_attr(
+        xml2::xml_children(xml2::xml_children(t[[i]])),
+        "TIME_SLOT_REF2"
+      )
+      a_id <- xml2::xml_attr(
+        xml2::xml_children(xml2::xml_children(t[[i]])),
+        "ANNOTATION_ID"
+      )
+      ar <- xml2::xml_attr(
+        xml2::xml_children(xml2::xml_children(t[[i]])),
+        "ANNOTATION_REF"
+      )
+      if (length(content) > 0) {
+        data.frame(
+          tier = i,
+          id = seq_along(content),
+          content = content,
+          tier_name = tier_names[i],
+          tier_type = tier_types[i],
+          ts_start = ts1,
+          ts_end = ts2,
+          a_id = a_id,
+          ar = ar,
+          stringsAsFactors = FALSE
+        )
       }
-    }) ->
-      r
+    })
+
     # merge list of dataframes  into dataframe
     r <- do.call(rbind, r)
 
[…]
