### 2020-09-27 55e385c
small code style changes and styler, 2. try
