+++ man/annotate_textgrid.Rd
@@ -36,12 +36,14 @@ Annotates textgrids. It is possible to define step in the argument "each",
 so each second element of the tier will be annotated.
 }
 \examples{
-annotate_textgrid(annotation = c("", "t", "e", "s", "t"),
-                  textgrid = system.file("extdata",
-                                         "test.TextGrid",
-                                         package = "phonfieldwork"),
-                  tier = 2, write = FALSE)
-
+annotate_textgrid(
+  annotation = c("", "t", "e", "s", "t"),
+  textgrid = system.file("extdata",
+    "test.TextGrid",
+    package = "phonfieldwork"
+  ),
+  tier = 2, write = FALSE
+)
 }
 \author{
 George Moroz <agricolamz@gmail.com>
