### 2020-09-28 58152fb
add create_empty_textgrid() function; correct docs
