+++ vignettes/phonfieldwork.Rmd
@@ -55,7 +55,6 @@ packageVersion("phonfieldwork")
 
 This tutorial can be cited as follows:
 
-
 ```r
 citation("phonfieldwork")
 ```
@@ -197,14 +196,14 @@ After collecting data and removing soundfiles with unsuccesful elicitations, one
 
 
 ```
-## │   ├── 2_01.wav
-## │   ├── 3_03.wav
-## │   └── logging.csv
+## ├── s1
+## │   ├── 01.wav
+## │   ├── 02.wav
+## │   └── 03.wav
 ## ├── s2
-## │   ├── 1_02.wav
-## │   ├── 2_01.wav
-## │   ├── 3_03.wav
-## │   └── logging.csv
+## │   ├── 01.wav
+## │   ├── 02.wav
+## │   └── 03.wav
 ```
 
 For each speaker `s1` and `s2` there is a folder that containes three audiofiles. Now let's rename the files.
@@ -217,29 +216,40 @@ rename_soundfiles(stimuli = my_stimuli_df$stimuli,
 ```
 
 ```
-## Error in rename_soundfiles(stimuli = my_stimuli_df$stimuli, prefix = "s1_", : There are some non-wav files:
-## logging.csv
+## You can find change correspondences in the following file:
+## /home/agricolamz/work/packages/phonfieldwork/vignettes/s1/backup/logging.csv
 ```
 
 As a result, you obtain the following structure:
 
 
 ```
-## ├── my_stimuli_df.xlsx
-## ├── phonfieldwork.Rmd
 ## ├── s1
-## │   ├── 1_02.wav
-## │   ├── 2_01.wav
-## │   ├── 3_03.wav
-## │   └── logging.csv
+## │   ├── 1_s1_tip.wav
+## │   ├── 2_s1_tap.wav
+## │   ├── 3_s1_top.wav
+## │   └── backup
+## │       ├── 01.wav
+## │       ├── 02.wav
+## │       ├── 03.wav
+## │       └── logging.csv
 ## ├── s2
-## │   ├── 1_02.wav
-## │   ├── 2_01.wav
-## │   ├── 3_03.wav
-## │   └── logging.csv
+## │   ├── 01.wav
+## │   ├── 02.wav
+## │   └── 03.wav
 ```
 
-The `rename_soundfiles()` function created a backup folder with all of the unrenamed files, and renamed all files using the prefix provided in the `prefix` argument. There is an additional argument `backup` that can be set to `FALSE` (it is `TRUE` by default), in case you are sure that the renaming function will work properly with your files and stimuli, and you do not need a backup of the unrenamed files.
[…]

[…]
