+++ tests/testthat/test-create-viewer.R
@@ -1,30 +1,10 @@
 test_that("create_viewer_general", {
   skip_on_cran()
   tmp <- tempdir()
-  dir.create(paste0(tmp, "/audio"))
-  download.file(
-    "https://github.com/agricolamz/phonfieldwork/raw/master/vignettes/s1/s1_sounds/1_s1_%C3%A6.wav",
-    paste0(tmp, "/audio/", "1_s1.wav")
-  )
-  download.file(
-    "https://github.com/agricolamz/phonfieldwork/raw/master/vignettes/s1/s1_sounds/2_s1_%C4%B1.wav",
-    paste0(tmp, "/audio/", "2_s1.wav")
-  )
-
-  dir.create(paste0(tmp, "/pic"))
-  download.file(
-    "https://github.com/agricolamz/phonfieldwork/raw/master/vignettes/s1/s1_pics/1_s1_%C3%A6.png",
-    paste0(tmp, "/pic/", "1_s1.png")
-  )
-  download.file(
-    "https://github.com/agricolamz/phonfieldwork/raw/master/vignettes/s1/s1_pics/2_s1_%C4%B1.png",
-    paste0(tmp, "/pic/", "2_s1.png")
-  )
-
   dir.create(paste0(tmp, "/result"))
   create_viewer(
-    audio_dir = paste0(tmp, "/audio"),
-    picture_dir = paste0(tmp, "/pic"),
+    audio_dir = "../audio",
+    picture_dir = "../pic",
     table = data.frame(id = 1:2),
     output_dir = paste0(tmp, "/result")
   )
@@ -39,7 +19,6 @@ test_that("create_viewer_general", {
   )
 
   test_that("create_viewer", {
-    skip_on_cran()
     expect_true(length(list.files(paste0(tmp, "/result"), "\\.html$")) == 2)
     expect_error(
       create_viewer(
@@ -59,7 +38,5 @@ test_that("create_viewer_general", {
   })
 
   unlink(paste0(tmp, "/result/"))
-  unlink(paste0(tmp, "/audio/"))
-  unlink(paste0(tmp, "/pic/"))
   rm(tmp)
 })
