### 2020-09-28 4a28678
add new vignette about data manipulation
