+++ vignettes/data_manipulation_with_tidyverse.Rmd
@@ -1,0 +1,227 @@
+---
+title: "Manipulating `phonfieldwork` data with `tidyverse`"
+author: "George Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
+date: "2020-09-29"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{Manipulating `phonfieldwork` data with `tidyverse`}
+  %\VignetteEncoding{UTF-8}
+  %\VignetteEngine{knitr::rmarkdown}
+editor_options: 
+  chunk_output_type: console
+---
+
+
+
+## Introduction
+
+The output tables from the `phonfieldwork`'s functions (e. g. `textgrid_to_df()`, `eaf_to_df()`, `exb_to_df()`, `flextext_to_df()` etc.) is hard to interpret since each row is a separate morpheme/observation or aother unit. In order to merge it to a more common representation we will use `tidyverse` functions (mainly `dplyr` and `tidyr` packages). This text will help you to achieve some results, but it is better to spend some times learning data manipulation with `dplyr` and `tidyr`.
+
+If you do not have `tidyverse` installed run:
+
+
+```r
+install.packages("tidyverse")
+```
+
+Let's load the package:
+
+
+```r
+library("tidyverse")
+```
+
+
+## .TextGrid, .eaf, .exb formats
+The standard sound annotation formats consisnt of tiers with parallel annotation:
+
+![](unnamed-chunk-31-1.png)
+
+If we convert this file to R we will achieve something like this:
+
+
+```r
+textgrid_to_df("s1/s1_all.TextGrid")
+#>    id time_start  time_end      content tier     tier_name          source
+#> 1   1  0.0000000 0.4821542          tip    1        labels s1_all.TextGrid
+#> 4   1  0.0000000 0.4821542 1_s1_tip.wav    2 backup labels s1_all.TextGrid
+#> 7   1  0.0000000 0.1063758                 3               s1_all.TextGrid
+#> 8   2  0.1063758 0.1843889            ı    3               s1_all.TextGrid
+#> 9   3  0.1843889 0.4821542                 3               s1_all.TextGrid
+#> 2   2  0.4821542 0.9120635          tap    1        labels s1_all.TextGrid
+#> 5   2  0.4821542 0.9120635 2_s1_tap.wav    2 backup labels s1_all.TextGrid
+#> 10  4  0.4821542 0.5805225                 3               s1_all.TextGrid
+#> 11  5  0.5805225 0.6784724            æ    3               s1_all.TextGrid
+#> 12  6  0.6784724 0.9120635                 3               s1_all.TextGrid
+#> 3   3  0.9120635 1.3942177          top    1        labels s1_all.TextGrid
+#> 6   3  0.9120635 1.3942177 3_s1_top.wav    2 backup labels s1_all.TextGrid
+#> 13  7  0.9120635 1.0390666                 3               s1_all.TextGrid
+#> 14  8  1.0390666 1.1300820            ɒ    3               s1_all.TextGrid
+#> 15  9  1.1300820 1.3942177                 3               s1_all.TextGrid
+```
+
[…]
