+++ vignettes/phonfieldwork.Rmd
@@ -2,7 +2,7 @@
 title: "Phonetic fieldwork and experiments with `phonfieldwork` package"
 author: "G. Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
 bibliography: bibliography.bib
-date: "2020-09-28"
+date: "2020-09-29"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Phonetic fieldwork and experiments with `phonfieldwork` package}
@@ -281,6 +281,15 @@ rename_soundfiles(stimuli = my_stimuli_df$stimuli,
 
 The last command renamed the soundfiles in the `s2` folder, adding the prefix `s2` as in the previous example, and the suffix `1`-`3`. On most operating systems it is impossible to create two files with the same name, so sometimes it can be useful to add some kind of index at the end of the files.
 
+There is also a possible scenario, that not all stimuli are retrieved from informant. So in order to deal with that case there is an additional argument `missing`, where user can put id numbers of stimuli that are not present in audiofiles:
+
+
+```r
+rename_soundfiles(stimuli = my_stimuli_df$stimuli,
+                  path = "s3/",
+                  missing = c(1, 3))
+```
+
 Sometimes it is useful to get information about sound duration:
 
 
@@ -342,7 +351,7 @@ This comand creates a new soundfile `s1_all.wav` and an asociated Praat TextGrid
 
 The resulting file can be parsed with Praat (subscripted *t* is the result of Praat's conversion):
 
-![](unnamed-chunk-22-1.png)
+![](unnamed-chunk-23-1.png)
 
 ## Annotate your data
 
@@ -362,7 +371,7 @@ annotate_textgrid(annotation =  my_stimuli_df$stimuli,
                   textgrid = "s1/s1_all.TextGrid")
 ```
 
-![](unnamed-chunk-24-1.png)
+![](unnamed-chunk-25-1.png)
 
 As you can see in the example, the `annotate_textgrid()` function creates a backup of the tier and adds a new tier on top of the previous one. It is possible to prevent the function from doing so by setting the `backup` argument to `FALSE`.
 
@@ -375,7 +384,7 @@ create_subannotation(textgrid = "s1/s1_all.TextGrid",
                      n_of_annotations = 3) # how many empty annotations per unit?
 ```
 
-![](unnamed-chunk-26-1.png)
+![](unnamed-chunk-27-1.png)
 
 Now we can annotate created tier:
 
@@ -387,13 +396,13 @@ annotate_textgrid(annotation = c("", "ı", "", "", "æ", "", "", "ɒ", ""),
                   backup = FALSE)
 ```
 
-![](unnamed-chunk-28-1.png)
+![](unnamed-chunk-29-1.png)
 
 You can see that we created a third tier with annotation. The only thing left is to move annotation boundaries in Praat (this can not be automated):
 
 
 
-![](unnamed-chunk-30-1.png)
+![](unnamed-chunk-31-1.png)
 
 You can see from the last figure that no backup tier was created (`backup = FALSE`), that the third tier was annotated (`tier = 3`).
 
@@ -403,7 +412,7 @@ In case you want to create an empty TextGrid it is possible to use a `create_emp
 ```r
 create_empty_textgrid("s2/s2_tip_1.wav", tiers = c("a", "b"))
 ```
-![](unnamed-chunk-32-1.png)
+![](unnamed-chunk-33-1.png)
 
 ```
 ## ├── s1
@@ -509,7 +518,7 @@ draw_sound("s1/s1_all.wav",
            to = 0.95)
 ```
 
-![](unnamed-chunk-38-1.png)
[…]
