### 2020-09-28 agricolamz/phonfieldwork#28 @agricolamz
.flextext without a word-level annotations
e. g. [this file](https://osf.io/xcafp/) can't be parsed because there is no any word-level annotations.

Thnks to Niko Partanen for spotting it.
