### 2020-09-29 2092200
change sections in reference; change function structures
