+++ tests/testthat/test-create-sound-play.R
@@ -1,3 +1,0 @@
-test_that("create_sound_play", {
-  expect_equal(nchar(create_sound_play("a.wav")), 117)
-})
