+++ tests/testthat/test-create-image-look-up.R
@@ -1,10 +1,0 @@
-test_that("create_image_look_up", {
-  expect_equal(nchar(create_image_look_up("a.png")), 121)
-  expect_error(
-    create_image_look_up("a.png", img_caption = 1:2),
-    paste0(
-      "It looks like the img_src variable contains 1 objects ",
-      "and the img_caption variable contains 2"
-    )
-  )
-})

[…]
