### 2020-09-29 agricolamz/phonfieldwork#30 @agricolamz
`create_viewer()` with part of images or sounds
Would it be possible to create a report where some sounds have images, but others don't? (Or even one where it's just sounds and no pictures, or just pictures and no sounds?)

Thanks to Jonathan Keane
