+++ presentations/2020.29.09_Conlab_phonfieldwork_talk/2020.29.09_Conlab_phonfieldwork_talk.Rmd
@@ -1,0 +1,436 @@
+---
+title: "Phonetic fieldwork and experiments with the `phonfieldwork` package for R"
+author: "George Moroz"
+institute: "Linguistic Convergence Laboratory, NRU HSE"
+date: |
+    | 28 September 2020
+    |
+    | Presentation is available here: \alert{tinyurl.com/y6lf5ch4}
+    | ![](images/01_qrcode.png)'
+output: 
+  beamer_presentation:
+    latex_engine: xelatex
+    citation_package: natbib
+    keep_tex: false
+    includes:
+      in_header: "config/presento.sty"
+bibliography: bibliography.bib
+biblio-style: "apalike"
+editor_options: 
+  chunk_output_type: console
+---
+
+```{r setup, include=FALSE}
+knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
+# library(qrcode)
+# png(filename="images/01_qrcode.png", width = 200, height = 200)
+# qrcode_gen("https://github.com/agricolamz/phonfieldwork/raw/master/presentations/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje_moroz_phonfieldwork.pdf")
+# dev.off()
+
+unlink("data", recursive = TRUE)
+dir.create("data")
+dir.create("data/s1")
+dir.create("data/s2")
+file.copy(paste0("../backup/s1/", list.files("../backup/s1/")), "data/s1/")
+file.copy(paste0("../backup/s2/", list.files("../backup/s2/")), "data/s2/")
+file.copy(paste0("../backup/", list.files("../backup/", pattern = "test")), "data/")
+```
+
+# Introduction
+
+## Most phonetic research consists of the following steps:
+
+1. Formulate a research question. Think of what kind of data is necessary to answer this question, what is the appropriate amount of data, what kind of annotation you will do, what kind of statistical models and visualizations you will use, etc.
+2. Create a list of stimuli.
+3. Elicite list of stimuli from speakers who signed an Informed Consent statement, agreeing to participate in the experiment to be recorded on audio and/or video.
+4. Annotate the collected data.
+5. Extract the information from annotated data.
+6. Create visualizations.
+7. Evaluate your statistical models.
+8. Report your results.
+9. Publish your data. \pause
+
+The `phonfieldwork` package is created for helping with items 3, 4 (partially), 5, 6, and 9.
+
+## Why/when do you need the phonfieldwork package?
+
+These ideal plan hides some technical subtasks:
+
+* creating a presentation for elicitation task
+* renaming and concatenating multiple sound files
+* automatic annotation in Praat TextGrids [@boersma19]
+* creating a searchable `.html` table with annotations, spectrograms and ability to hear sound
+* converting multiple formats (Praat, ELAN [@brugman04] and EXMARaLDA [@schmidt09]) \pause
+
+All of these tasks can be solved by a mixture of different tools:
+
+* any programming language can handle automatic file renaming
+* Praat contains scripts for concatenating and renaming files \pause
+
+`phonfieldwork` provides a functionality that will make it easier to solve those tasks independently of any additional tools. You can also compare the functionality with other packages: `rPraat` [@borzil16], `textgRid` [@reidy16], `pympi` [@lubbers13]
+
+## Philosophy of the `phonfieldwork` package
+
+* each stimulus is a separate file
+* researcher carefully listens to consultants to make sure that they are producing the kind of speech they wanted
[…]

[…]
