+++ presentations/2020.29.09_Conlab_phonfieldwork_talk/bibliography.bib
@@ -6,6 +6,14 @@
   year={2003}
 }
 
+@inproceedings{draxler04,
+  title={SpeechRecorder-a Universal Platform Independent Multi-Channel Audio Recording Software.},
+  author={Draxler, Christoph and J{\"a}nsch, Klaus},
+  booktitle={LREC},
+  year={2004}
+}
+
+
 @misc{lubbers13,
 	author={Lubbers, Mart and Torreira, Francisco},
 	year={2013},
@@ -45,7 +53,7 @@
     note = {R package version 1.0.1},
     url = {https://CRAN.R-project.org/package=textgRid},
   }
-  
+
 @inproceedings{brugman04,
   title={{A}nnotating {M}ulti-media/{M}ulti-modal {R}esources with {ELAN}.},
   author={Brugman, Hennie and Russel, Albert and Nijmegen, Xd},
@@ -53,7 +61,7 @@
   year={2004}
 }
 
-  
+
 @article{schmidt09,
   title={EXMARaLDA--Creating, analysing and sharing spoken language corpora for pragmatic research},
   author={Schmidt, Thomas and W{\"o}rner, Kai},
@@ -65,8 +73,8 @@
   publisher={John Benjamins}
 }
 
-  
-  	
+
+
   @InProceedings{borzil16,
     author = {Tomáš Bořil and Radek Skarnitzl},
     editor = {Petr Sojka and Aleš Horák and Ivan Kopeček and Karel Pala},
