+++ presentations/2020.29.09_Conlab_phonfieldwork_talk/2020.29.09_Conlab_phonfieldwork_talk.Rmd
@@ -5,8 +5,8 @@ institute: "Linguistic Convergence Laboratory, NRU HSE"
 date: |
     | 28 September 2020
     |
-    | Presentation is available here: \alert{tinyurl.com/y6lf5ch4}
-    | ![](images/01_qrcode.png)'
+    | Presentation is available here: \alert{tinyurl.com/y2x8ppnl}
+    | ![](images/00_qrcode.png)'
 output: 
   beamer_presentation:
     latex_engine: xelatex
@@ -23,8 +23,8 @@ editor_options:
 ```{r setup, include=FALSE}
 knitr::opts_chunk$set(echo = TRUE, warning = FALSE, message = FALSE)
 # library(qrcode)
-# png(filename="images/01_qrcode.png", width = 200, height = 200)
-# qrcode_gen("https://github.com/agricolamz/phonfieldwork/raw/master/presentations/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje/2020.08.08_Grupo_de_estad%C3%ADstica_para_el_estudio_del_lenguaje_moroz_phonfieldwork.pdf")
+# png(filename="images/00_qrcode.png", width = 200, height = 200)
+# qrcode_gen("https://github.com/agricolamz/phonfieldwork/raw/master/presentations/2020.29.09_Conlab_phonfieldwork_talk/2020.29.09_Conlab_phonfieldwork_talk.pdf")
 # dev.off()
 
 unlink("data", recursive = TRUE)
@@ -34,6 +34,34 @@ dir.create("data/s2")
 file.copy(paste0("../backup/s1/", list.files("../backup/s1/")), "data/s1/")
 file.copy(paste0("../backup/s2/", list.files("../backup/s2/")), "data/s2/")
 file.copy(paste0("../backup/", list.files("../backup/", pattern = "test")), "data/")
+```
+
+# About `phonfieldwork`
+
+## About `phonfieldwork`
+
+This package was started as a help for our guest student Margaux Dubuis from Switzerland, who was going to study New Caledonian language Vamale (`vama1243`).
+
+```{r, echo=FALSE, out.width = '60%'}
+knitr::include_graphics("images/01_Margaux.jpg")
+```
+
+## About `phonfieldwork`
+
+* 2019-08-24 --- first release on CRAN
+* Now there is a v0.0.7 on CRAN, and total 6363 downloads
+
+```{r, echo=FALSE, out.width = '100%'}
+knitr::include_graphics("images/02_commits.png")
+```
+
+## About `phonfieldwork`
+
+* On June 20 I have made [a submission](https://github.com/ropensci/software-review/issues/385) to [rOpenSci](https://ropensci.org/).
+* Here are my reviewers:
+
+```{r, echo=FALSE, out.width = '90%'}
+knitr::include_graphics("images/03_Niko_Jonathan.png")
 ```
 
 # Introduction
@@ -60,7 +88,7 @@ These ideal plan hides some technical subtasks:
 * renaming and concatenating multiple sound files
 * automatic annotation in Praat TextGrids [@boersma19]
 * creating a searchable `.html` table with annotations, spectrograms and ability to hear sound
-* converting multiple formats (Praat, ELAN [@brugman04] and EXMARaLDA [@schmidt09]) \pause
+* converting multiple formats (Praat, ELAN [@brugman04] and EXMARaLDA [@schmidt09], .srt and some others) \pause
 
 All of these tasks can be solved by a mixture of different tools:
 
@@ -75,7 +103,7 @@ All of these tasks can be solved by a mixture of different tools:
 * researcher carefully listens to consultants to make sure that they are producing the kind of speech they wanted
[…]
