+++ R/concatenate_soundfiles.R
@@ -1,14 +1,12 @@
 #' Concatenate sounds
 #'
-#' Creates a sound file and a Praat TextGrid whose interval labels are the
-#' original names of the sound
+#' Creates a merged sound file from old sound files in a folder. If the annotation argument is not equal to \code{NULL}, it creates an annotation file (Praat .TextGrid, ELAN .eaf or EXMARaLDA .exb) with original sound names annotation.
 #'
 #' @author George Moroz <agricolamz@gmail.com>
 #'
 #' @param path path to the directory with soundfiles.
 #' @param result_file_name name of the result and annotation files.
-#' @param annotation there are several variants: "textgrid" for Praat TextGrid,
-#' "eaf" for ELAN's .eaf file, or "exb" for EXMARaLDA's .exb file
+#' @param annotation character. There are several variants: "textgrid" for Praat TextGrid, "eaf" for ELAN's .eaf file, or "exb" for EXMARaLDA's .exb file. It is also possible to use \code{NULL} in order to prevent the creation of the annotation file.
 #' @return
 #'
 #' @examples
@@ -20,7 +18,7 @@
 #'
 #' # here are two .wav files in a folder
 #' list.files(tdir)
-#' # [1] "post.wav" "test.wav"
+#' # [1] "post.wav" "test.wav" ...
 #'
 #' # Concatenate all files from the folder into concatenated.wav and create
 #' # corresponding TextGrid
@@ -28,6 +26,7 @@
 #'
 #' list.files(tdir)
 #' # [1] "concatenated.TextGrid" "concatenated.wav" "post.wav" "test.wav" ...
+#'
 #' @return no output
 #' @export
 #' @importFrom tuneR readWave
@@ -40,9 +39,9 @@
 concatenate_soundfiles <- function(path,
                                    result_file_name = "concatenated",
                                    annotation = "textgrid") {
+  match.arg(annotation, c("textgrid", "eaf", "exb"))
 
   # concatenate sounds ------------------------------------------------------
-
   files <- list.files(
     path = normalizePath(path),
     pattern = "(\\.WAVE?$)|(\\.wave?$)|(\\.MP3?$)|(\\.mp3?$)"
@@ -89,62 +88,58 @@ concatenate_soundfiles <- function(path,
   }
   sound <- do.call(tuneR::bind, list)
   tuneR::writeWave(sound, paste0(path, "/", result_file_name, ".wav"))
-  # create a TextGrid -------------------------------------------------------
 
-  if (annotation == "textgrid") {
-    duration <- unlist(lapply(list, function(i) {
-      length(i@left) / i@samp.rate
-    }))
-    start_time <- c(0, cumsum(duration[-length(duration)]))
-    end_time <- cumsum(duration)
-    my_textgrid <- data.frame(
-      TierNumber = 1,
-      TierName = "annotation",
-      TierType = "IntervalTier",
-      Index = seq_along(files),
-      StartTime = start_time,
-      EndTime = end_time,
-      Label = sort(files),
-      stringsAsFactors = FALSE
-    )
-    writeLines(
-      c(
-        'File type = "ooTextFile"',
-        'Object class = "TextGrid"',
-        "",
-        "xmin = 0 ",
-        paste0("xmax = ", end_time[length(end_time)]),
-        "tiers? <exists> ",
-        "size = 1 ",
-        "item []: ",
-        "    item [1]:",
-        '        class = "IntervalTier"',
-        '        name = "labels"',
-        "        xmin = 0",
-        paste0("        xmax = ", end_time[length(end_time)]),
-        paste0("        intervals: size = ", length(duration)),
-        paste0(paste0("        intervals [", my_textgrid$Index, "]:"),
-          "\n",
[…]
