+++ man/create_empty_textgrid.Rd
@@ -4,14 +4,24 @@
 \alias{create_empty_textgrid}
 \title{Create an empty TextGrid}
 \usage{
-create_empty_textgrid(file_name, tiers = NULL, point_tiers = NULL)
+create_empty_textgrid(
+  duration,
+  tiers = NULL,
+  point_tiers = NULL,
+  path,
+  result_file_name = "new_textgrid"
+)
 }
 \arguments{
-\item{file_name}{a reference sound file.}
+\item{duration}{integer. Duration of the textgrid. If you do not know the duration of your audio file use the \code{get_sound_duration()} function.}
 
 \item{tiers}{a vector that contain tier names.}
 
 \item{point_tiers}{a vector that defines which tiers should be made point tiers. This argument excepts numeric values (e. g. \code{c(2, 4)} means second and forth tiers) or character (e. g. \code{c("a", "b")} means tiers with names "a" and "b")}
+
+\item{path}{path to the directory with soundfiles.}
+
+\item{result_file_name}{name of the result and annotation files.}
 }
 \value{
 The function returns no output, just creates a Praat TextGrid in the same folder as a reference sound file.
