### 2020-10-03 935de4a
update create_glossed_document()
