+++ NEWS.md
@@ -8,7 +8,7 @@
 - add an argument `external` to the `create_presentation()` function in order to mark external images or gifs
 - remove all `encoding` arguments and replace it with encoding autodetection from `uchardet` (thanks to Artem Klevtsov for help)
 - add `autonumber`, `loging` and `missing` arguments to `rename_soundfiles()` function (thanks to Niko Partanen)
-- a lot of minor style changes (tnaks to Jonathan Keane)
+- a lot of minor style changes (thanks to Jonathan Keane)
 - add the `create_empty_textgrid()` function (thanks to Niko Partanen)
 - add the `data_manipulation_with_tidyverse` vignette (thanks to Niko Partanen)
 - add the `concatenate_textgrids()` function
