### 2020-10-03 0eda420
update create_glossed_document()
