### 2020-10-03 e4e9818
update create_glossed_document()
