+++ DESCRIPTION
@@ -27,11 +27,10 @@ RoxygenNote: 7.1.1
 VignetteBuilder: knitr
 Suggests: 
     knitr,
-    testthat,
-    covr,
     dplyr,
     tidyr,
     DT,
-    lingtypology
+    lingtypology,
+    testthat
 Language: en-US
 Config/testthat/parallel: true
