### 2020-10-05 759aa22
add read_from_folder(); remove _from_folder arguments
