+++ tests/testthat/test-read-from-folder.R
@@ -1,0 +1,6 @@
+textgrids <- read_from_folder(system.file("extdata", package = "phonfieldwork"),
+                              textgrid_to_df)
+
+test_that("read_from_folder", {
+  expect_true(nrow(textgrids) == 42)
+})

[…]
