### 2020-10-05 92a56cb
add read_from_folder(); remove _from_folder arguments
