+++ man/get_sound_duration.Rd
@@ -14,7 +14,7 @@ Calculate sound(s) duration.
 }
 \examples{
 get_sound_duration(
-  sounds_from_folder = system.file("extdata", package = "phonfieldwork")
+  system.file("extdata", "test.wav", package = "phonfieldwork")
 )
 }
 \author{
