### 2020-10-05 #385 @agricolamz
So now in order to separate functions for reading individual files and reading all the files in the directory I created a function `read_from_folder <- function(path, FUN){...}` which reads each file format from folder according to some existing function in `phonfieldwork`, e. g. `read_from_folder("test", textgrid_to_df)`. What do you think, should I also create a separate functions like `textgrid_to_df_from_folder()`, `eaf_to_df_from_folder()`?
