### 2020-10-17 #385 @jonkeane
Hello again,

I’ve reviewed the changes and the responses and I’m satisfied with both.

For the encoding issues, I was using some of the files I found in http://www.ims.uni-stuttgart.de/phonetik/helps/praat-scripting/scripts.zip that gave me some of the encoding issues I mentioned.
