### 2020-10-19 #385 @melvidoni
Since both reviewers approved... Approved! Thanks @agricolamz  for submitting and @jonkeane @nikopartanen  for your reviews!

To-dos:
- [ ] Transfer the repo to rOpenSci's "ropensci" GitHub organization under "Settings" in your repo.  I have invited you to a team that should allow you to do so.  You'll be made admin once you do. Tag me here so I can give you access.
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
