+++ DESCRIPTION
@@ -13,14 +13,30 @@ Imports:
     xml2,
     uchardet,
     tools
-Authors@R: c(person("George", "Moroz",
-             role = c("aut", "cre"),
-             email = "agricolamz@gmail.com",
-             comment = c(ORCID="0000-0003-1990-6083")))
+Authors@R: c(
+  person(
+    given = "George", 
+    family = "Moroz",
+    role = c("aut", "cre"),
+    email = "agricolamz@gmail.com",
+    comment = c(ORCID="0000-0003-1990-6083")),
+  person(
+    given = "Jonathan", 
+    family = "Keane",
+    role = c("rev"),
+    email = "jkeane@gmail.com",
+    comment = c(ORCID = "0000-0001-7087-9776")),
+  person(
+    given = "Niko", 
+    family = "Partanen",
+    role = c("rev"),
+    email = "nikotapiopartanen@gmail.com",
+    comment = c(ORCID = "0000-0001-8584-3880"))
+  )
 Description: There are a lot of different typical tasks that have to be solved during phonetic research and experiments. This includes creating a presentation that will contain all stimuli, renaming and concatenating multiple sound files recorded during a session, automatic annotation in 'Praat' TextGrids (this is one of the sound annotation standards provided by 'Praat' software, see Boersma & Weenink 2018 <http://www.fon.hum.uva.nl/praat/>), creating an html table with annotations and spectrograms, and converting multiple formats ('Praat' TextGrid, 'ELAN', 'EXMARaLDA', 'Audacity', subtitles '.srt', and 'FLEx' flextext). All of these tasks can be solved by a mixture of different tools (any programming language has programs for automatic renaming, and Praat contains scripts for concatenating and renaming files, etc.). 'phonfieldwork' provides a functionality that will make it easier to solve those tasks independently of any additional tools. You can also compare the functionality with other packages: 'rPraat' <https://CRAN.R-project.org/package=rPraat>, 'textgRid' <https://CRAN.R-project.org/package=textgRid>.
 License: GPL (>= 2)
-URL: https://CRAN.R-project.org/package=phonfieldwork, https://agricolamz.github.io/phonfieldwork/
-BugReports: https://github.com/agricolamz/phonfieldwork/issues
+URL: https://CRAN.R-project.org/package=phonfieldwork, https://ropensci.github.io/phonfieldwork/
+BugReports: https://github.com/ropensci/phonfieldwork/issues
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
