+++ vignettes/phonfieldwork.Rmd
@@ -2,7 +2,7 @@
 title: "Phonetic fieldwork and experiments with `phonfieldwork` package"
 author: "G. Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
 bibliography: bibliography.bib
-date: "2020-10-05"
+date: "2020-10-21"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Phonetic fieldwork and experiments with `phonfieldwork` package}
@@ -32,10 +32,10 @@ This command will install the last stable version of the `phonfieldwork` package
 
 ```r
 install.packages("remotes")
-remotes::install_github("agricolamz/phonfieldwork")
+remotes::install_github("ropensci/phonfieldwork")
 ```
 
-If you have any trouble installing the package, you will not be able to use its functionality. In that case you can create [an issue on Github](https://github.com/agricolamz/phonfieldwork/issues) or send an email. Since this package could completely destroy your data, **please do not use it until you are sure that you have made a backup**.
+If you have any trouble installing the package, you will not be able to use its functionality. In that case you can create [an issue on Github](https://github.com/ropensci/phonfieldwork/issues) or send an email. Since this package could completely destroy your data, **please do not use it until you are sure that you have made a backup**.
 
 Use the `library()` command to load the package:
 
@@ -62,8 +62,8 @@ citation("phonfieldwork")
 
 ```
 ## 
-## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork
-## package_. <URL: https://CRAN.R-project.org/package=phonfieldwork>.
+## Moroz G (2020). _Phonetic fieldwork and experiments with phonfieldwork package_. <URL:
+## https://CRAN.R-project.org/package=phonfieldwork>.
 ## 
 ## A BibTeX entry for LaTeX users is
 ## 
@@ -75,7 +75,7 @@ citation("phonfieldwork")
 ##   }
 ```
 
-If you have any trouble using the package, do not hesitate to create [an issue on Github](https://github.com/agricolamz/phonfieldwork/issues/new).
+If you have any trouble using the package, do not hesitate to create [an issue on Github](https://github.com/ropensci/phonfieldwork/issues/new).
 
 
 # Philosophy of the `phonfieldwork` package
@@ -102,7 +102,7 @@ There are some phoneticians who prefer to record everything, for language docume
 
 You can show a native speaker your stimuli one by one or not show them the stimule but ask them to pronounce a certain stimulus or its translation. I use presentations to collect all stimuli in a particular order without the risk of omissions.
 
[…]
