+++ vignettes/data_manipulation_with_tidyverse.Rmd
@@ -1,7 +1,7 @@
 ---
 title: "Manipulating `phonfieldwork` data with `tidyverse`"
 author: "George Moroz, [NRU HSE Linguistic Convergence Laboratory](https://ilcl.hse.ru/en/)"
-date: "2020-10-05"
+date: "2020-10-21"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Manipulating `phonfieldwork` data with `tidyverse`}
@@ -84,41 +84,34 @@ Imagine that we obtained the first result from `flextext_to_df()`:
 ```r
 df <- flextext_to_df("files/zilo_test.flextext")
 head(df)
-#>   p_id s_id w_id    txt     cf hn     gls                   msa
-#> 1    1    1    1     б-     б-  1      an Inflects any category
-#> 2    1    1    1    ик1    ик1  1    быть                    гл
-#> 3    1    1    1     -о     -о  1     pst               гл:Past
-#> 4    1    1    1     -й     -й  5 cvb(pf)    гл:Converb/Perfect
-#> 5    1    1    1 =гъоди =гъоди  1    =rep                  част
-#> 6    1    1    2     б-     б-  1      an Inflects any category
-#>                  free_trans                          text_title
-#> 1 Жил-был (у Гъули?) петух. 2017.04 Fairytale about the rooster
-#> 2 Жил-был (у Гъули?) петух. 2017.04 Fairytale about the rooster
-#> 3 Жил-был (у Гъули?) петух. 2017.04 Fairytale about the rooster
-#> 4 Жил-был (у Гъули?) петух. 2017.04 Fairytale about the rooster
-#> 5 Жил-был (у Гъули?) петух. 2017.04 Fairytale about the rooster
-#> 6 Жил-был (у Гъули?) петух. 2017.04 Fairytale about the rooster
-#>                                  morph                                 word
-#> 1 d7f713db-e8cf-11d3-9764-00c04f186933 efafb420-e203-4685-9be2-1b7810f10a70
-#> 2 d7f713e8-e8cf-11d3-9764-00c04f186933 efafb420-e203-4685-9be2-1b7810f10a70
-#> 3 d7f713dd-e8cf-11d3-9764-00c04f186933 efafb420-e203-4685-9be2-1b7810f10a70
-#> 4 d7f713dd-e8cf-11d3-9764-00c04f186933 efafb420-e203-4685-9be2-1b7810f10a70
-#> 5 d7f713e1-e8cf-11d3-9764-00c04f186933 efafb420-e203-4685-9be2-1b7810f10a70
-#> 6 d7f713db-e8cf-11d3-9764-00c04f186933 c76d26b7-b84a-42a8-ba34-38e712b1db13
-#>                                 phrase                            paragraph
-#> 1 1cbadc4f-4051-4783-a0d8-bfeee2d2fb13 0c9ffe63-b4bf-4af3-a1da-f68567e03513
-#> 2 1cbadc4f-4051-4783-a0d8-bfeee2d2fb13 0c9ffe63-b4bf-4af3-a1da-f68567e03513
-#> 3 1cbadc4f-4051-4783-a0d8-bfeee2d2fb13 0c9ffe63-b4bf-4af3-a1da-f68567e03513
-#> 4 1cbadc4f-4051-4783-a0d8-bfeee2d2fb13 0c9ffe63-b4bf-4af3-a1da-f68567e03513
-#> 5 1cbadc4f-4051-4783-a0d8-bfeee2d2fb13 0c9ffe63-b4bf-4af3-a1da-f68567e03513
-#> 6 1cbadc4f-4051-4783-a0d8-bfeee2d2fb13 0c9ffe63-b4bf-4af3-a1da-f68567e03513
[…]
