### 2020-12-06 #385 @agricolamz
Dear @stefaniebutland, I would like to make a blogpost about phonfieldwork. May I get a publication date?
