### 2020-12-07 #385 @stefaniebutland
Excellent @agricolamz. There are a couple of options for publication dates, depending on how urgent it is for you.

1. publish Tues Jan 19; submit draft by Tues Jan 12.
2. publish Tues Dec 22; submit draft by Mon Dec 14, with agreement to have all edits done before Fri Dec 18

Blog post drafts are reviewed by rOpenSci Community Assistant @steffilazerte. Since your [lingtypology post](https://ropensci.org/blog/2017/05/16/lingtypology/) we have written https://blogguide.ropensci.org/ with detailed content and technical guidelines.
